Hierarchical clustering of topography metrics.

This function adds the column cluster to objects of class fgeo_topography. Using the stats package, add_cluster() calculates a hierarchical clustering of the columns meanelev, slope and convexity. First, it calculates a dissimilarities among those columns (with stats::dist() and all its defaults), then it calculates a tree (with stats::hclust() and all its defaults), and finally cuts the tree in n groups (with stats::cutree()).

add_cluster(.data, n)

Arguments

.data
A dataframe of subclass fgeo_topography.

n
Integer. Number of cluster-groups to construct (passed to the argument k to stats::cutree()).

Examples

elev_ls <- fgeo.data::luquillo_elevation
topo <- fgeo_topography(elev_ls, gridsize = 20)
add_cluster(topo, n = 4)

#> # A tibble: 400 x 6
#>       gx    gy meanelev convex slope cluster
#>  * <dbl> <dbl>    <dbl>  <dbl> <dbl>   <int>
#>  1     0     0     365. -0.390  8.87       1
#>  2     0    20     365. -0.285 13.7        1
#>  3     0    40     367.  0.210 14.4        1
#>  4     0    60     367.  0.34  15.9        1
#>  5     0    80     362. -0.255 20.8        1
#>  6     0   100     358. -0.510 12.4        1
#>  7     0   120     359. -0.28  17.1        1
#>  8     0   140     360. -0.265 22.1        1
#>  9     0   160     362.  0.055 24.3        1
#> 10     0   180     364.  0.415 15.2        1
#> # ... with 390 more rows