Add columns to a dataframe.

These functions wrap legacy code from the CTFS RPackage.

add_var(x, var, gridsize = 20, plotdim = NULL)

add_lxly(x, gridsize = 20, plotdim = NULL)

add_index(x, gridsize = 20, plotdim = NULL)

add_rowcol(x, gridsize = 20, plotdim = NULL)

add_hectindex(x, gridsize = 20, plotdim = NULL)

add_quad(x, gridsize = 20, plotdim = NULL, start = 1, width = 2)

Arguments

x
A ForestGEO table, either a ViewFullTable or a census table.

var
A character string; either "lxly", "rowcol", "index", hectindex.

gridsize
The gridsize of the census plot (commonly 20 m).

plotdim
The global dimensions of the census plot (i.e. the maximum possible values of `gx` and `gy`).

start
1 or 0, indicating how to label the first plot-column.

width
Number; width to padd the labels of plot-columns and -rows (passed to stringr::str_pad()).

Value

A modified version of the dataframe x with the additional variable(s) var.

See also

stringr::str_pad().

Other functions to add columns to dataframes.: add_status_tree

Examples

x <- tibble::tibble(gx = c(0, 50, 999.9, 1000), gy = gx/2)

# Each `add_var(x, var = "*")` has a shortcut: `add_*()`
add_var(x, var = "lxly")

#> Gessing: plotdim = c(1000, 500)
#>   * If this guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 4
#>       gx    gy    lx    ly
#>    <dbl> <dbl> <dbl> <dbl>
#> 1    0     0     0    0   
#> 2   50.0  25.0  10.0  5.00
#> 3 1000   500    19.9 19.9 
#> 4 1000   500    NA   NA   

# same
add_lxly(x)

#> Gessing: plotdim = c(1000, 500)
#>   * If this guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 4
#>       gx    gy    lx    ly
#>    <dbl> <dbl> <dbl> <dbl>
#> 1    0     0     0    0   
#> 2   50.0  25.0  10.0  5.00
#> 3 1000   500    19.9 19.9 
#> 4 1000   500    NA   NA   

add_index(x)

#> Gessing: plotdim = c(1000, 500)
#>   * If this guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 3
#>       gx    gy   index
#>    <dbl> <dbl>   <dbl>
#> 1    0     0      1.00
#> 2   50.0  25.0   52.0 
#> 3 1000   500   1250   
#> 4 1000   500     NA   

add_rowcol(x)

#> Gessing: plotdim = c(1000, 500)
#>   * If this guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 4
#>       gx    gy   row   col
#>    <dbl> <dbl> <dbl> <dbl>
#> 1    0     0    1.00  1.00
#> 2   50.0  25.0  2.00  3.00
#> 3 1000   500   25.0  50.0 
#> 4 1000   500   NA    NA   

add_hectindex(x)

#> Gessing: plotdim = c(1000, 500)
#>   * If this guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 3
#>       gx    gy hectindex
#>    <dbl> <dbl>     <dbl>
#> 1    0     0        1.00
#> 2   50.0  25.0      1.00
#> 3 1000   500       50.0 
#> 4 1000   500       56.0 

add_quad(x)

#> Gessing: plotdim = c(1000, 500)
#>   * If this guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 3
#>       gx    gy quad 
#>    <dbl> <dbl> <chr>
#> 1    0     0   0101 
#> 2   50.0  25.0 0302 
#> 3 1000   500   5025 
#> 4 1000   500   NANA 

add_quad(x, start = 0)

#> Gessing: plotdim = c(1000, 500)
#>   * If this guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 3
#>       gx    gy quad 
#>    <dbl> <dbl> <chr>
#> 1    0     0   0000 
#> 2   50.0  25.0 0201 
#> 3 1000   500   4924 
#> 4 1000   500   NANA 

# `width` gives the nuber of digits to pad the label of plot-rows and
# plot-columns, e.g. 3 pads plot-rows with three zeros and plot-columns with an
# extra trhree zeros, resulting in a total of 6 zeros.
add_quad(x, start = 0, width = 3)

#> Gessing: plotdim = c(1000, 500)
#>   * If this guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 3
#>       gx    gy quad  
#>    <dbl> <dbl> <chr> 
#> 1    0     0   000000
#> 2   50.0  25.0 002001
#> 3 1000   500   049024
#> 4 1000   500   NANA