Make any function work with grouped data.

Make any function work with grouped data.

by_group(.x, .f, ...)

Arguments

.x
A dataframe.

.f
A function.

...
Additional arguments passed to .f.

Value

The value of .f applied to groups.

Acknowledgments

Tristan Mahr and Edward Visel helped improve this function (post).

Examples

dfm <- tibble::tribble(
  ~x, ~y,  ~z,
  11, 21,  31,
  12, 22,  32,
  13, 23,  33
)

# Say you want to make this funciton work with grouped data
first_row <- function(.x, to_chr = FALSE) {
  first <- .x[1, ]
  if (to_chr) {
    first[] <- lapply(first, as.character)
  }

  tibble::as.tibble(first)
}

library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following object is masked from 'package:testthat':
#> 
#>     matches

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

# Ungrouped
dfm %>% first_row()

#> # A tibble: 1 x 3
#>       x     y     z
#>   <dbl> <dbl> <dbl>
#> 1    11    21    31

# Same -- it does nothing with ungrouped data
dfm %>% by_group(first_row)

#> # A tibble: 1 x 3
#>       x     y     z
#>   <dbl> <dbl> <dbl>
#> 1    11    21    31

# Grouped
dfm %>% group_by(x) %>% by_group(first_row)

#> # A tibble: 3 x 3
#> # Groups:   x [3]
#>       x     y     z
#>   <dbl> <dbl> <dbl>
#> 1    11    21    31
#> 2    12    22    32
#> 3    13    23    33

dfm %>% group_by(x) %>% by_group(first_row, to_chr = TRUE)

#> # A tibble: 3 x 3
#> # Groups:   x [3]
#>   x     y     z    
#>   <chr> <chr> <chr>
#> 1 11    21    31   
#> 2 12    22    32   
#> 3 13    23    33