Collapse the values of CensusID.

This function is useful to summarize the status history of a tree tag across censuses. It is particularly useful to avoid duplicated tree tags on maps.

collapse_censusid(x)

Arguments

x
A dataframe -- specifically a ViewFullTable.

Value

A modified version of x, most likely with less rows.

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

x <- dplyr::tribble(
  ~CensusID, ~Tag, ~Status,
        "1",  "1", "alive",
        "1",  "1",  "dead",
        "1",  "2",  "dead",
        "1",  "2",  "dead",
        "2",  "1", "alive",
        "2",  "1", "alive",
        "2",  "2", "alive",
        "2",  "2",  "dead"
)

collapse_censusid(x)

#> # A tibble: 4 x 3
#>   CensusID Tag   Status
#>   <chr>    <chr> <chr> 
#> 1 1, 2     1     alive 
#> 2 1, 2     1     dead  
#> 3 1, 2     2     dead  
#> 4 1, 2     2     alive 

with_status_tree <- fgeo.tool::add_status_tree(x, "alive", "dead")
collapsed_cns <- collapse_censusid(with_status_tree)
unique(select(collapsed_cns, -Status))

#> # A tibble: 3 x 3
#>   CensusID Tag   status_tree
#>   <chr>    <chr> <chr>      
#> 1 1, 2     1     alive      
#> 2 1, 2     2     dead       
#> 3 1, 2     2     alive