Import multiple files (.csv, excel, .Rdata, ...) from a directory into a list.

These functions read from a specific directory where every file has its own extension as indicated by each function's name. Notice that function names have the format input_output, i.e. file-extension_list. If none of these functions do what you want, create your own with read_with().

rdata_list(path_dir, ...)

rda_list(path_dir, ...)

rds_list(path_dir, ...)

csv_list(path_dir, ...)

delim_list(path_dir, ...)

tsv_list(path_dir, ...)

Arguments

path_dir
String; the path to a directory containing the files to read (all must be of appropriate format; see examples).

...
Arguments passed to the reader function.

Value

A list of dataframes.

See also

Examples

path_rdata <- tool_example("rdata")
path_rdata

#> [1] "C:/Users/LeporeM/Documents/Dropbox/git/ml/fgeo.tool/inst/extdata/rdata"

dir(path_rdata)

#> [1] "tree5.RData" "tree6.RData"

rdata_list(path_rdata)

#> $tree5
#> # A tibble: 3 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1    104    143 10009 10009   DACE~ 113      10.3  245.    439947        5 193  
#> 2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    466597        5  40.4
#> 3    180    225 1001~ 100174  CASA~ 921     165.   410.    466623        5  45  
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#> 
#> $tree6
#> # A tibble: 3 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1    104    143 10009 10009   DACE~ 113      10.3  245.    582850        6 195  
#> 2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    578696        6  44.9
#> 3    180    225 1001~ 100174  CASA~ 921     165.   410.    617049        6  46.1
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#> 

path_rds <- tool_example("rds")
path_rds

#> [1] "C:/Users/LeporeM/Documents/Dropbox/git/ml/fgeo.tool/inst/extdata/rds"

dir(path_rds)

#> [1] "tree5.rds" "tree6.rds"

rds_list(path_rds)

#> $tree5
#> # A tibble: 3 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1    104    143 10009 10009   DACE~ 113      10.3  245.    439947        5 193  
#> 2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    466597        5  40.4
#> 3    180    225 1001~ 100174  CASA~ 921     165.   410.    466623        5  45  
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#> 
#> $tree6
#> # A tibble: 3 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1    104    143 10009 10009   DACE~ 113      10.3  245.    582850        6 195  
#> 2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    578696        6  44.9
#> 3    180    225 1001~ 100174  CASA~ 921     165.   410.    617049        6  46.1
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#> 

path_csv <- tool_example("csv")
path_csv

#> [1] "C:/Users/LeporeM/Documents/Dropbox/git/ml/fgeo.tool/inst/extdata/csv"

dir(path_csv)

#> [1] "csv1.csv" "csv2.csv"

csv_list(path_csv)

#> Parsed with column specification:
#> cols(
#>   numbers = col_double(),
#>   letters = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   numbers = col_double(),
#>   letters = col_character()
#> )

#> $csv1
#> # A tibble: 3 x 2
#>   numbers letters
#>     <dbl> <chr>  
#> 1       1 a      
#> 2       2 b      
#> 3       3 c      
#> 
#> $csv2
#> # A tibble: 3 x 2
#>   numbers letters
#>     <dbl> <chr>  
#> 1       4 a      
#> 2       5 b      
#> 3       6 c      
#> 

path_tsv <- tool_example("tsv")
path_tsv

#> [1] "C:/Users/LeporeM/Documents/Dropbox/git/ml/fgeo.tool/inst/extdata/tsv"

dir(path_tsv)

#> [1] "tsv1.tsv" "tsv2.tsv"

tsv_list(path_tsv)

#> Parsed with column specification:
#> cols(
#>   numbers = col_double(),
#>   letters = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   numbers = col_double(),
#>   letters = col_character()
#> )

#> $tsv1
#> # A tibble: 3 x 2
#>   numbers letters
#>     <dbl> <chr>  
#> 1       1 a      
#> 2       2 b      
#> 3       3 c      
#> 
#> $tsv2
#> # A tibble: 3 x 2
#>   numbers letters
#>     <dbl> <chr>  
#> 1       4 a      
#> 2       5 b      
#> 3       6 c      
#> 

# Weird: Tab separated columns in a file with .csv extension
path_weird <- tool_example("weird")
dir(path_weird)

#> [1] "tsv1.csv" "tsv2.csv"

# Extension is .csv, but this is not what you want
csv_list(path_weird)

#> Parsed with column specification:
#> cols(
#>   `numbers	letters` = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   `numbers	letters` = col_character()
#> )

#> $tsv1
#> # A tibble: 3 x 1
#>   `numbers\tletters`
#>   <chr>             
#> 1 "1\ta"            
#> 2 "2\tb"            
#> 3 "3\tc"            
#> 
#> $tsv2
#> # A tibble: 3 x 1
#>   `numbers\tletters`
#>   <chr>             
#> 1 "4\ta"            
#> 2 "5\tb"            
#> 3 "6\tc"            
#> 

# Use this instead
delim_list(path_weird, delim = "\t")

#> Parsed with column specification:
#> cols(
#>   numbers = col_double(),
#>   letters = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   numbers = col_double(),
#>   letters = col_character()
#> )

#> $tsv1
#> # A tibble: 3 x 2
#>   numbers letters
#>     <dbl> <chr>  
#> 1       1 a      
#> 2       2 b      
#> 3       3 c      
#> 
#> $tsv2
#> # A tibble: 3 x 2
#>   numbers letters
#>     <dbl> <chr>  
#> 1       4 a      
#> 2       5 b      
#> 3       6 c      
#>