fgeo.tool: Import and Manipulate ForestGEO Data

To help you access, transform, analyze, and visualize ForestGEO data, we developed a collection of R packages (<https://forestgeo.github.io/fgeo/>). This package, in particular, helps you to easily import, filter, and modify ForestGEO data. To learn more about ForestGEO visit <http://www.forestgeo.si.edu/>.

See also