Construct habitat data.

This function constructs an object of class fgeo_habitat. It can handle a multiple inputs and always outputs the same data structure.

fgeo_habitat(elevation, gridsize, n, ...)

# S3 method for list
fgeo_habitat(elevation, gridsize, n, ...)

# S3 method for data.frame
fgeo_habitat(elevation, gridsize, n, xdim, ydim, ...)

Arguments

elevation

A list with at least three elements: col containing elevation data; and xdim and ydim giving plot dimensions; OR

A dataframe containing elevation data, in which case the parameters xdim and ydim must be provided. It may be the element col of a ForestGEO elevation-list or an object of class fgeo_elevation (see fgeo_elevation()).

gridsize
Number to round x and y by. Commonly, gridsize = 20.

n
Number of elevation groups (habitats) to cut elevation data by. Commonly, n = 4.

...
Other arguments passed to methods.

xdim, ydim
x and y dimensions of the plot.

Value

A dataframe of habitat data.

Examples

# Input: Object of class list
elev_list <- fgeo.data::luquillo_elevation
hab1 <- fgeo_habitat(elev_list, gridsize = 20, n = 4)
str(hab1)

#> Classes ‘fgeo_habitat’, ‘tbl_df’, ‘tbl’ and 'data.frame':	6400 obs. of  3 variables:
#>  $ gx      : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ gy      : int  0 5 10 15 20 25 30 35 40 45 ...
#>  $ habitats: int  1 1 1 1 1 1 1 1 1 1 ...

# Input: Object of class dataframe
elev_df <- fgeo.data::luquillo_elevation$col
hab2 <- fgeo_habitat(elev_df, gridsize = 20, n = 4, xdim = 320, ydim = 500)
str(hab2)

#> Classes ‘fgeo_habitat’, ‘tbl_df’, ‘tbl’ and 'data.frame':	6400 obs. of  3 variables:
#>  $ gx      : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ gy      : int  0 5 10 15 20 25 30 35 40 45 ...
#>  $ habitats: int  1 1 1 1 1 1 1 1 1 1 ...

# Input: Object of class fgeo_elev
fgeo_elev <- fgeo_elevation(fgeo.data::luquillo_elevation)
hab3 <- fgeo_habitat(elev_df, gridsize = 20, n = 4, xdim = 320, ydim = 500)
str(hab3)

#> Classes ‘fgeo_habitat’, ‘tbl_df’, ‘tbl’ and 'data.frame':	6400 obs. of  3 variables:
#>  $ gx      : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ gy      : int  0 5 10 15 20 25 30 35 40 45 ...
#>  $ habitats: int  1 1 1 1 1 1 1 1 1 1 ...

identical(hab1, hab2)

#> [1] TRUE

identical(hab2, hab3)

#> [1] TRUE

hab1

#> # A tibble: 6,400 x 3
#>       gx    gy habitats
#>  * <int> <int>    <int>
#>  1     0     0        1
#>  2     0     5        1
#>  3     0    10        1
#>  4     0    15        1
#>  5     0    20        1
#>  6     0    25        1
#>  7     0    30        1
#>  8     0    35        1
#>  9     0    40        1
#> 10     0    45        1
#> # ... with 6,390 more rows