Calculate mean elevation, convexity and slope.

Calculate mean elevation, convexity and slope.

fgeo_topography(elevation, ...)

# S3 method for data.frame
fgeo_topography(elevation, gridsize, xdim = NULL,
  ydim = NULL, edgecorrect = TRUE, ...)

# S3 method for list
fgeo_topography(elevation, gridsize, edgecorrect = TRUE,
  ...)

Arguments

elevation
One of these:

A list with at least three elements: col containing elevation data; and xdim and ydim giving plot dimensions; OR

A dataframe containing elevation data, in which case the parameters xdim and ydim must be provided. It may be the element col of a ForestGEO elevation-list or an object of class fgeo_elevation (see fgeo_elevation()).

...
Other arguments passed to methods.

gridsize
Number giving the size of each quadrat for which a habitat is calculated. Commonly, gridsize = 20.

xdim, ydim
(If elevation is a dataframe) x and y dimensions of the plot.

edgecorrect
Correct convexity in edge quadrats?

Value

A dataframe of subclass fgeo_topography.

Input

The main input can be either the elevation list that ForestGEO delivers, or the element col of such list -- which is a dataframe containing the elevation data. Notice that the required arguments vary according to the main input (the elevation list or the elevation dataframe). Whatever the input, the dataframe containing the elevation data must have columns gx and gy or x and y.

See also

Examples

elev_list <- fgeo.data::luquillo_elevation
fgeo_topography(elev_list, gridsize = 20)

#> # A tibble: 400 x 5
#>       gx    gy meanelev convex slope
#>  * <dbl> <dbl>    <dbl>  <dbl> <dbl>
#>  1     0     0     365. -0.390  8.87
#>  2     0    20     365. -0.285 13.7 
#>  3     0    40     367.  0.210 14.4 
#>  4     0    60     367.  0.34  15.9 
#>  5     0    80     362. -0.255 20.8 
#>  6     0   100     358. -0.510 12.4 
#>  7     0   120     359. -0.28  17.1 
#>  8     0   140     360. -0.265 22.1 
#>  9     0   160     362.  0.055 24.3 
#> 10     0   180     364.  0.415 15.2 
#> # ... with 390 more rows

elev_df <- elev_list$col
fgeo_topography(elev_df, gridsize = 20, xdim = 320, ydim = 500)

#> # A tibble: 400 x 5
#>       gx    gy meanelev convex slope
#>  * <dbl> <dbl>    <dbl>  <dbl> <dbl>
#>  1     0     0     365. -0.390  8.87
#>  2     0    20     365. -0.285 13.7 
#>  3     0    40     367.  0.210 14.4 
#>  4     0    60     367.  0.34  15.9 
#>  5     0    80     362. -0.255 20.8 
#>  6     0   100     358. -0.510 12.4 
#>  7     0   120     359. -0.28  17.1 
#>  8     0   140     360. -0.265 22.1 
#>  9     0   160     362.  0.055 24.3 
#> 10     0   180     364.  0.415 15.2 
#> # ... with 390 more rows