Create field forms forms.

Create field forms. Data should be from a single plot and of a single census. Pandoc must be installed (see section Installing Pandoc).

fieldforms_output(forms, filename = "fieldforms.docx",
  output_format = "word_document", header = fieldforms_header(),
  output_dir = getwd())

Arguments

forms
A list of dataframes -- most likely the output of fieldforms_prepare().

filename
String; path to the output file. Must include the extension ".docx". Passed to rmarkdown::render().

output_format
String. Currently only "word_document" is supported.

header
A string to use as the header of the fieldform of each quadrat.

output_dir
String giving a directory to save the file in.

Value

A (MS Word) .docx file.

Installing Pandoc

This function requires that you have pandoc installed. See installation instructions at https://pandoc.org/installing.html. To install pandoc from R see the function install.pandoc() of the installr package (https://goo.gl/cpaxtd).

See also

fieldforms_header().

Examples

# NOT RUN {
library(fgeo.tool)
x <- top4quad
prep <- fieldforms_prepare(x)
lapply(prep, head)

# You may customize the prepared data before making the fielforms. For example:
# Making names uppercase
prep <- lapply(prep, function(x) setNames(x, toupper(names(x))))
# Renaming a column
prep <- lapply(prep, function(x) dplyr::rename(x, COMMENTS = COMM))
# Removing a column
prep <- lapply(prep, function(x) dplyr::select(x, -NOTE))
# Adding a column
prep <- lapply(prep, function(x) dplyr::mutate(x, NEW = " "))

# Now you can make the fieldforms
tmp <- tempdir()
fieldforms_output(prep, output_dir = tmp)
fieldforms_output(
  prep, "fieldforms2.docx", header = "Custom Header", output_dir = tmp
)

# Confirm
dir(tmp)[grepl("docx$", dir(tmp))]

# Cleaning temporary directory
unlink(tmp)
# }