Read all .csv or excel files from a directory into dataframes.

Read all .csv, .xls, or .xlsx files from a directory into a dataframe or a list of dataframes, where each element of the list is named as the source file.

csv_to_df(input_dir)

csv_to_df_lst(input_dir)

xl_to_df(input_dir)

xl_to_df_lst(input_dir)

Arguments

input_dir
String giving the directory containing the excel workbooks to read from.

Value

The versions ending in _df output a single dataframe. The versions ending in _lst output a list of dataframes.

Acknowledgment

Thanks to Jessica Shue for inspiring this functions.

Examples

path <- system.file("extdata", "files/01.csv", package = "fgeo.tool")
input_dir <- fs::path_dir(path)
input_dir

#> C:/Users/LeporeM/Dropbox/git_repos/fgeo.tool/inst/extdata/files

dir(input_dir)

#> [1] "01.csv"  "01.xls"  "02.csv"  "02.xlsx"

# Read all .csv files
csv_to_df(input_dir)

#> Parsed with column specification:
#> cols(
#>   x = col_integer(),
#>   y = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   x = col_integer(),
#>   y = col_character()
#> )

#> # A tibble: 10 x 2
#>        x y    
#>    <int> <chr>
#>  1     1 a    
#>  2     2 b    
#>  3     3 c    
#>  4     4 d    
#>  5     5 e    
#>  6     1 k    
#>  7     2 l    
#>  8     3 m    
#>  9     4 n    
#> 10     5 o    

csv_to_df_lst(input_dir)

#> Parsed with column specification:
#> cols(
#>   x = col_integer(),
#>   y = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   x = col_integer(),
#>   y = col_character()
#> )

#> $`01.csv`
#> # A tibble: 5 x 2
#>       x y    
#>   <int> <chr>
#> 1     1 a    
#> 2     2 b    
#> 3     3 c    
#> 4     4 d    
#> 5     5 e    
#> 
#> $`02.csv`
#> # A tibble: 5 x 2
#>       x y    
#>   <int> <chr>
#> 1     1 k    
#> 2     2 l    
#> 3     3 m    
#> 4     4 n    
#> 5     5 o    
#> 

# Read all .xls or .xlsx files
xl_to_df_lst(input_dir)

#> $`01.xls`
#> # A tibble: 5 x 2
#>       x y    
#>   <dbl> <chr>
#> 1     1 a    
#> 2     2 b    
#> 3     3 c    
#> 4     4 d    
#> 5     5 e    
#> 
#> $`02.xlsx`
#> # A tibble: 5 x 2
#>       x y    
#>   <dbl> <chr>
#> 1     1 k    
#> 2     2 l    
#> 3     3 m    
#> 4     4 n    
#> 5     5 o    
#> 

xl_to_df(input_dir)

#> # A tibble: 10 x 2
#>        x y    
#>    <dbl> <chr>
#>  1     1 a    
#>  2     2 b    
#>  3     3 c    
#>  4     4 d    
#>  5     5 e    
#>  6     1 k    
#>  7     2 l    
#>  8     3 m    
#>  9     4 n    
#> 10     5 o