Filter a (fgeo) dataframe by the status of each stem or tree.

In stem- and ViewFull-tables status and Status refer to each stem. In tree-tables status refers to tree. To get the status of each tree based on the status of its stems see add_status_tree().

filter_status(x, wood, .status, exclude = FALSE)

drop_dead_stem(x, .status = "D")

drop_dead_tree(x, .status = "D")

Arguments

x
A ForestGEO table, either a ViewFullTable or a census table.

wood
Either "stem" or "tree", to indicate if the data should be filtered based on the status of each individual stem or tree (notice that one tree is dead only when not some but all its stems are dead).

.status
Character vector; Must be one of possible values of the variable giving the status of the dataframe x.

exclude
Logical; TRUE filters the data to exclude the values passed to .status.

Value

A filtered version of the dataframe x.

See also

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(fgeo.tool)

# Filter by the status of each stem (wood = "stem") -----------------------

# CENSUS TABLE: STEM TABLE

x <- bciex::bci12s7mini
table(x$status)

#> 
#>    A    D    G 
#>  665  434 1066 

result <- filter_status(x, wood = "stem", .status = "D")
table(result$status)

#> 
#>   D 
#> 434 

result <- filter_status(x, wood = "stem", .status = "D", exclude = TRUE)
table(result$status)

#> 
#>    A    G 
#>  665 1066 

# Shortcut
result <- drop_dead_stem(x)
table(result$status)

#> 
#>    A    G 
#>  665 1066 

# Warns
result <- filter_status(x, wood = "stem", .status = c("A", "wrong-status"))

#> Warning: No observation has .status = wrong-status
#>   * Detected values: A, G, D

table(result$status)

#> 
#>   A 
#> 665 

# CENSUS TABLE: TREE TABLE

# Works exactly in the same way
x <- bciex::bci12t7mini
table(x$status)

#> 
#>   A   D 
#> 577 423 

result <- filter_status(x, wood = "stem", .status = "D")
table(result$status)

#> 
#>   D 
#> 423 

# Shortcut
result <- drop_dead_stem(x)
table(result$status)

#> 
#>   A 
#> 577 

# VIEWFULL TABLE

# Works exactly in the same way, but notice the following:
# * The variable Status starts with capital S;
# * The values of Status are not, say "A" or "D", but "alive" or "dead".
x <- bciex::bci12vft_mini
table(x$Status)

#> 
#> alive  dead 
#>  3853   521 

result <- filter_status(x, wood = "stem", .status = "alive")
table(result$Status)

#> 
#> alive 
#>  3853 

# Warns because `.status` defaults to "D" -- not to "dead".
result <- drop_dead_stem(x)

#> Warning: No observation has .status = D
#>   * Detected values: alive, dead

# Fix and repeat
result <- drop_dead_stem(x, .status = "dead")
table(result$Status)

#> 
#> alive 
#>  3853 

# Filter by the status of each tree (wood = "tree") -----------------------

# CENSUS TABLE: STEM TABLE

x <- bciex::bci12s7mini

# Add the variable status_tree, which gives the status of each tree, not stem
unique(x$status)

#> [1] "A" "G" "D"

x <- add_status_tree(x, status_d = "D", status_a = "A")
table(x$status_tree)

#> 
#>    A    D 
#> 1809  356 

result <- filter_status(x, wood = "tree", .status = "A")
table(result$status_tree)

#> 
#>    A 
#> 1809 

result <- filter_status(x, wood = "tree", .status = "D", exclude = TRUE)
table(result$status_tree)

#> 
#>    A 
#> 1809 

# Shortcut
result <- drop_dead_tree(x)
result %>%
  dplyr::arrange(tag, stemID, status) %>%
  dplyr::select(tag, stemID, status, status_tree)

#> # A tibble: 1,809 x 4
#>    tag    stemID status status_tree
#>    <chr>   <int> <chr>  <chr>      
#>  1 000851      1 A      A          
#>  2 001080      1 G      A          
#>  3 001080      2 G      A          
#>  4 001080      3 G      A          
#>  5 001080      4 G      A          
#>  6 001080      5 G      A          
#>  7 001080      6 G      A          
#>  8 001080      7 G      A          
#>  9 001080      8 G      A          
#> 10 001080      9 G      A          
#> # ... with 1,799 more rows

table(result$status_tree)

#> 
#>    A 
#> 1809 

# CENSUS TABLE: TREE TABLE
x <- bciex::bci12t7mini
# For a tree census-table, each stem maps to a tree, so the value of the
# variable `status` gives, at the same time, the status of the stem and tree.
x <- add_status_tree(x, "D", "A")
identical(x$status, x$status_tree)

#> [1] TRUE

# So the result will be the same if we use `wood = tree` or `wood = stem`.
result1 <- filter_status(x, wood = "tree", .status = "A")
result2 <- filter_status(x, wood = "stem", .status = "A")
identical(result1, result2)

#> [1] TRUE

# Shortcut
result <- drop_dead_tree(x, .status = "D")
identical(result, result1)

#> [1] TRUE

# VIEWFULL TABLE

x <- bciex::bci12vft_mini

# Add the variable status_tree, which gives the status of each tree, not stem
unique(x$Status)

#> [1] "alive" "dead" 

x <- add_status_tree(x, status_d = "dead", status_a = "alive")
table(x$status_tree)

#> 
#> alive  dead 
#>  3853   521 

result <- filter_status(x, wood = "tree", .status = "alive")
table(result$status_tree)

#> 
#> alive 
#>  3853 

# Shortcut
unique(x$Status)

#> [1] "alive" "dead" 

result <- drop_dead_tree(x, "dead")
result %>%
  dplyr::arrange(Tag, StemID, Status) %>%
  dplyr::select(Tag, StemID, Status, status_tree)

#> # A tibble: 3,853 x 4
#>    Tag    StemID Status status_tree
#>    <chr>   <int> <chr>  <chr>      
#>  1 000002      1 alive  alive      
#>  2 000002      1 alive  alive      
#>  3 000004      1 alive  alive      
#>  4 000004      1 alive  alive      
#>  5 000006      7 alive  alive      
#>  6 000006      7 alive  alive      
#>  7 000007      1 alive  alive      
#>  8 000007      1 alive  alive      
#>  9 000008      1 alive  alive      
#> 10 000008      1 alive  alive      
#> # ... with 3,843 more rows

table(result$status_tree)

#> 
#> alive 
#>  3853