Functions to get variables from other variables.

These functions wrap their corresponding functions from the CTFS R Package, but these versions are stricter. The main differences are these:

names use "_" not ".".

argument gridsize defaults to missing to force the user to provide it.

If the argument plotdim is missing from functions gxgy_fun(), its value will be guessed and notified.

rowcol_to_index(rowno, colno, gridsize, plotdim)

index_to_rowcol(index, gridsize, plotdim)

gxgy_to_index(gx, gy, gridsize, plotdim)

gxgy_to_lxly(gx, gy, gridsize, plotdim)

gxgy_to_qxqy(gx, gy, gridsize, plotdim)

gxgy_to_rowcol(gx, gy, gridsize, plotdim)

gxgy_to_hectindex(gx, gy, plotdim)

index_to_gxgy(index, gridsize, plotdim)

Arguments

rowno, colno
Row and column number -- as defined in a census plot.

gridsize
The gridsize of the census plot (commonly 20 m).

plotdim
The global dimensions of the census plot (i.e. the maximum possible values of gx and gy).

index
Index number as defined for a census plot.

gx, gy
A number; global x and y position in a census plot.

Value

A vector or dataframe (see examples).

Details

gxgy_to_qxqy() didn't exist in the original CTFS R Package. Added for consistency.

Author

Rick Condit, Suzanne Lao.

Examples

gxgy_to_index(c(0, 400, 990), c(0, 200, 490), gridsize = 20)

#> Guessing: plotdim = c(1000, 500)

#> [1]    1  511 1250

gridsize <- 20
plotdim <- c(1000, 500)

x <- gxgy_to_hectindex(1:3, 1:3, plotdim)
x

#> [1] 1 1 1

typeof(x)

#> [1] "double"

is.data.frame(x)

#> [1] FALSE

is.vector(x)

#> [1] TRUE

x <- gxgy_to_index(1:3, 1:3, gridsize, plotdim)
x

#> [1] 1 1 1

typeof(x)

#> [1] "double"

is.data.frame(x)

#> [1] FALSE

is.vector(x)

#> [1] TRUE

x <- gxgy_to_lxly(1:3, 1:3, gridsize, plotdim)
x

#>   lx ly
#> 1  1  1
#> 2  2  2
#> 3  3  3

typeof(x)

#> [1] "list"

is.data.frame(x)

#> [1] TRUE

is.vector(x)

#> [1] FALSE

x <- gxgy_to_rowcol(1:3, 1:3, gridsize, plotdim)
x

#>   row col
#> 1   1   1
#> 2   1   1
#> 3   1   1

typeof(x)

#> [1] "list"

is.data.frame(x)

#> [1] TRUE

is.vector(x)

#> [1] FALSE

x <- index_to_rowcol(1:3, gridsize, plotdim)
x

#>   row col
#> 1   1   1
#> 2   2   1
#> 3   3   1

typeof(x)

#> [1] "list"

is.data.frame(x)

#> [1] TRUE

is.vector(x)

#> [1] FALSE

x <- rowcol_to_index(1:3, 1:3, gridsize, plotdim)
x

#> [1]  1 27 53

typeof(x)

#> [1] "double"

is.data.frame(x)

#> [1] FALSE

is.vector(x)

#> [1] TRUE

index_to_gxgy(1:3, gridsize, plotdim)

#>   gx gy
#> 1  0  0
#> 2  0 20
#> 3  0 40