Replace the string 'NULL' by a missing character value (NA_character_)

Importing a dataset can be challenging. A common problem occurrs when reading data a dataset with missing values encoded as 'NULL'. The problem is this: When you read that dataset, R will interpret those missing values incorrectly -- not as NA but as the literal string 'NULL'. This funciton sanitizes a dataset by converting the literal string 'NULL' to the missing charecter value NA_character_.

literal_null_to_na(x)

Arguments

x
A dataframe or vector.

Value

A modified version of x with literals "NULL" replaced by NA_character_.

Details

Instead of using this function, you should avoid missinterpreting 'NULL' in the first place. You can pass 'NULL' to the argument na of the functions read_*() of the readr package (e.g. read_csv()). For example:

readr::read_csv("your-file.csv", na = c("", "NA", "NULL"))

For more details on how to read data simply and safely see this blog post.

Examples

chr <- c("a", "NULL")
literal_null_to_na(chr)

#> [1] "a" NA 

dfm <- data.frame(
  a = c(1, 2, "NULL"), b = c(1, "NULL", 3),
  stringsAsFactors = FALSE
)
dfm

#>      a    b
#> 1    1    1
#> 2    2 NULL
#> 3 NULL    3

literal_null_to_na(dfm)

#> # A tibble: 3 x 2
#>   a     b    
#>   <chr> <chr>
#> 1 1     1    
#> 2 2     <NA> 
#> 3 <NA>  3