Lower and restore names.

These functions are useful together, to lowercase names, then to do something (as long as attributes are preserved -- see section warning), and then restore the original names.

nms_lowercase(x)

nms_restore(x)

Arguments

x
A named object.

Value

nms_lowercase() Returns the object x with lowercase names

nms_restore() Returns the object x with original (restored) names.

Warning

nms_restore() is similar to nms_restore_newvar() but nms_restore() is necessary if the data is mutated with dplyr::mutate(): dplyr::mutate() drops attributes (https://github.com/tidyverse/dplyr/issues/1984), which makes it nms_restore() useless. attributes.

See also

Examples

cns <- tibble::tibble(CensusID = 1, status = "A")
original <- cns
original

#> # A tibble: 1 x 2
#>   CensusID status
#>      <dbl> <chr> 
#> 1        1 A     

lowered <- nms_lowercase(cns)
lowered

#> # A tibble: 1 x 2
#>   censusid status
#>      <dbl> <chr> 
#> 1        1 A     

attr(lowered, "names_old")

#> [1] "CensusID" "status"  

back_to_original <- nms_restore(lowered)
back_to_original

#> # A tibble: 1 x 2
#>   CensusID status
#>      <dbl> <chr> 
#> 1        1 A