Restore the names of a dataframe to which a new variable has been added.

This function helps to develop functions that work with both ViewFullTables and census tables. ViewFullTables and census tables share multiple names but often the case of those names is different (e.g. Tag and tag). When developing functions that work with both ViewFullTables and fgeo tables, one solution is to lowercase all names, do whatever the function needs to do, and then restore the old names. This function helps by restoring old names, which is not straight forward when the function adds a new variable and may contain a preexisting variable with the same name of the added variable.

nms_restore_newvar(x, new_var, old_nms)

Arguments

x
A dataframe.

new_var
The name of a single new variable added to x.

old_nms
A vector containing the old names of x.

Value

Returns the input with the names changed accordingly.

Details

The length of x must equal the number of names in old_nms, or that + 1".

nms_restore_newvar() is similar to nms_restore() but specifically targets dataframes that have been mutated with dplyr::mutate(). dplyr::mutate() drops attributes (https://github.com/tidyverse/dplyr/issues/1984), which makes it nms_restore() useless. attributes.

See also

Examples

# Data does not contain the variable that will be added
dfm <- data.frame(X = 1, Y = "a")
(old <- names(dfm))

#> [1] "X" "Y"

# Lower names
(dfm <- rlang::set_names(dfm, tolower))

#>   x y
#> 1 1 a

# Add a variable
mutated <- dplyr::mutate(dfm, newvar = x + 1)
# Restore
nms_restore_newvar(mutated, "newvar", old)

#>   X Y newvar
#> 1 1 a      2

# Data contains the variable that will be added
dfm <- data.frame(X = 1, Y = "a", newvar = "2")
(old <- names(dfm))

#> [1] "X"      "Y"      "newvar"

# Lower names
(dfm <- rlang::set_names(dfm, tolower))

#>   x y newvar
#> 1 1 a      2

# Add a variable
mutated <- dplyr::mutate(dfm, newvar = x + 1)
# Restore
nms_restore_newvar(mutated, "newvar", old)

#>   X Y newvar
#> 1 1 a      2