Create tidy names that are lowercase and have no empty spaces.
Source: R/nms_manipulate_names.R

These functions create tidy names that are lowercase and have no empty spaces. nms_tidy() tidyes the names of named objects. Unnamed strings are also tidied via str_as_tidy_names(). str_as_tidy_names() tidies a string -- not its names.

nms_tidy(x)

Arguments

x
A named object or a character string.

Value

A modified version of x with tidy names or a string of tidy names.

Examples

messy <- "Hi yOu"

# With unnamed strings, both functions do the same
unnamed_string <- messy
names(messy)

#> NULL

nms_tidy(messy)

#> [1] "hi_you"

# Same
str_as_tidy_names(messy)

#> [1] "hi_you"

# WHY TWO FUNCTIONS?

messy_named_string <- c(`Messy Name` = messy)
# Targets names
nms_tidy(messy_named_string)

#> messy_name 
#>   "Hi yOu" 

# Targets strings -- not its names
str_as_tidy_names(messy_named_string)

#> Messy Name 
#>   "hi_you" 

# Same output, but here `str_as_tidy_names()` better communicates intention.
dfm <- data.frame(1)
setNames(dfm, nms_tidy(messy))

#>   hi_you
#> 1      1

setNames(dfm, str_as_tidy_names(messy))

#>   hi_you
#> 1      1

# Makes more sense when operating on strings
setNames(list(1), str_as_tidy_names(messy))

#> $hi_you
#> [1] 1
#> 

# Makes more sense when operating on named objects
messy_list <- list(`Hi yOu` = 1)
nms_tidy(messy_list)

#> $hi_you
#> [1] 1
#>