Pick rows from ForestGEO censuses in a list.

This function allows you to pick rows from a key census (dataframe) in a list and pick the exact same rows in all other non-key censuses.

pick(.data, ..., key = NULL)

Arguments

.data
A ForestGEO dataset of class 'censuses_lst' (a list).

...
Other arguments passed to methods.

key
Key dataframe to pick rows from and recycle in all other censuses.

See also

Examples

censuses <- as_censuses(
  list(
    c1 = tibble(dbh = 1:2),
    c2 = tibble(dbh = 8:9)
  )
)
class(censuses)

#> [1] "censuses_lst" "list"        

censuses

#> $c1
#> # A tibble: 2 x 1
#>     dbh
#>   <int>
#> 1     1
#> 2     2
#> 
#> $c2
#> # A tibble: 2 x 1
#>     dbh
#>   <int>
#> 1     8
#> 2     9
#> 

pick(censuses, dbh == 1, key = "c1")

#> $c1
#> # A tibble: 1 x 1
#>     dbh
#>   <int>
#> 1     1
#> 
#> $c2
#> # A tibble: 1 x 1
#>     dbh
#>   <int>
#> 1     8
#> 

# Same because `key` defaults to the first element of the list
pick(censuses, dbh == 1)

#> $c1
#> # A tibble: 1 x 1
#>     dbh
#>   <int>
#> 1     1
#> 
#> $c2
#> # A tibble: 1 x 1
#>     dbh
#>   <int>
#> 1     8
#> 

# Different `key`
pick(censuses, dbh == 1, key = "c2")

#> $c1
#> # A tibble: 0 x 1
#> # ... with 1 variable: dbh <int>
#> 
#> $c2
#> # A tibble: 0 x 1
#> # ... with 1 variable: dbh <int>
#> 

# With multiple conditions passed via `...`
pick(censuses, dbh >= 1, dbh < 2)

#> $c1
#> # A tibble: 1 x 1
#>     dbh
#>   <int>
#> 1     1
#> 
#> $c2
#> # A tibble: 1 x 1
#>     dbh
#>   <int>
#> 1     8
#> 

# `pick()` is a common step after reading multiple censuses into a list
rdata_files <- tool_example("rdata")
dir(rdata_files)

#> [1] "tree5.RData" "tree6.RData"

censuses <- as_censuses(rdata_list(rdata_files))
pick(censuses, dbh >= 100)

#> $tree5
#> # A tibble: 1 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1    104    143 10009 10009   DACE~ 113      10.3  245.    439947        5   193
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#> 
#> $tree6
#> # A tibble: 1 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1    104    143 10009 10009   DACE~ 113      10.3  245.    582850        6   195
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#> 

# Same as a pipeline
rdata_files %>%
  rdata_list() %>%
  as_censuses() %>%
  pick(dbh >= 100)

#> $tree5
#> # A tibble: 1 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1    104    143 10009 10009   DACE~ 113      10.3  245.    439947        5   193
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#> 
#> $tree6
#> # A tibble: 1 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1    104    143 10009 10009   DACE~ 113      10.3  245.    582850        6   195
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#>