Pick woods after collapsing multiple stems by treeid and censusid.

These functions pick rows by groups. At the core they run dplyr::filter() but include these additional features:

They reject multiple plots with an informative error message.

They operate within groups define with dplyr::group_by().

If the data has multiple censuses, they automatically group by census.

They collapse data of multi-stem trees by picking a single stem per treeid. The picked stem depends on the values of hom or dbh detected for each census of each treeid. The selection is as follows:

If there is a single hom value, it is the stem of maximum dbh.

If there is more than one hom value, it is the stem of maximum dbh.

pick_woods() is a general function that takes any dataframe and any number of expressions to filter the dataframe.

pick_trees() picks stems of 100 mm dbh and above.

pick_saplings() picks stems between 10 mm dbh inclusive and 100 mm dbh exclusive.

pick_woods(.data, ...)

pick_trees(.data)

pick_saplings(.data)

Arguments

.data
A ForestGEO-like dataframe, either a census or ViewFullTable.

...
Expressions passed to dplyr::filter(), E.g. the expression dbh >= 100 picks dbh values of 100 mm and above.

Value

A dataframe. See description.

Examples

census <- tibble::tribble(
    ~sp, ~treeID, ~stemID,  ~hom, ~dbh,
  "sp1",     "1",   "1.2",   130,  122,
  "sp1",     "1",   "1.1",   130,   10,
  "sp2",     "2",   "2.1",   130,   22,
  "sp2",     "2",   "2.2",   130,   99,
  "sp2",     "2",   "2.2",   144,   88,
  "sp2",     "2",   "2.3",   130,   NA,
)

# Piks largest hom first, then largest dbh (to correct effect of batreesses)
pick_woods(census, dbh >=10, dbh < 100)

#> # A tibble: 1 x 5
#>   sp    treeID stemID   hom   dbh
#>   <chr> <chr>  <chr>  <dbl> <dbl>
#> 1 sp2   2      2.2      144    88

# Same
pick_saplings(census)

#> # A tibble: 1 x 5
#>   sp    treeID stemID   hom   dbh
#>   <chr> <chr>  <chr>  <dbl> <dbl>
#> 1 sp2   2      2.2      144    88

pick_trees(census)

#> # A tibble: 1 x 5
#>   sp    treeID stemID   hom   dbh
#>   <chr> <chr>  <chr>  <dbl> <dbl>
#> 1 sp1   1      1.2      130   122