Print all rows and/or columns of a tibble.

Print all rows and/or columns of a tibble.

# S3 method for tbl
print(x, ..., n = NULL, width = NULL, n_extra = NULL)

# S3 method for tbl_df
print(x, ..., n = NULL, width = NULL,
  n_extra = NULL)

Arguments

x
Object to show.

...
Other arguments passed to format().

n
Number of rows to show. If NULL, the default, will print all rows if less than option tibble.print_max. Otherwise, will print tibble.print_min rows.

width
Width of text output to generate. This defaults to NULL, which means use getOption("tibble.width") or (if also NULL) getOption("width"); the latter displays only the columns that fit on one screen. You can also set options(tibble.width = Inf) to override this default and always print all columns.

n_extra
Number of extra columns to print abbreviated information for, if the width is too small for the entire tibble. If NULL, the default, will print information about at most tibble.max_extra_cols extra columns.

Value

Print output (and the main input invisibly).

Examples

print(as_tibble(iris), n = 1)

#> # A tibble: 150 x 5
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>          <dbl>       <dbl>        <dbl>       <dbl> <fct>  
#> 1          5.1         3.5          1.4         0.2 setosa 
#> # ... with 149 more rows

print(as_tibble(iris), n = 100)

#> # A tibble: 150 x 5
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width Species   
#>            <dbl>       <dbl>        <dbl>       <dbl> <fct>     
#>   1          5.1         3.5          1.4         0.2 setosa    
#>   2          4.9         3            1.4         0.2 setosa    
#>   3          4.7         3.2          1.3         0.2 setosa    
#>   4          4.6         3.1          1.5         0.2 setosa    
#>   5          5           3.6          1.4         0.2 setosa    
#>   6          5.4         3.9          1.7         0.4 setosa    
#>   7          4.6         3.4          1.4         0.3 setosa    
#>   8          5           3.4          1.5         0.2 setosa    
#>   9          4.4         2.9          1.4         0.2 setosa    
#>  10          4.9         3.1          1.5         0.1 setosa    
#>  11          5.4         3.7          1.5         0.2 setosa    
#>  12          4.8         3.4          1.6         0.2 setosa    
#>  13          4.8         3            1.4         0.1 setosa    
#>  14          4.3         3            1.1         0.1 setosa    
#>  15          5.8         4            1.2         0.2 setosa    
#>  16          5.7         4.4          1.5         0.4 setosa    
#>  17          5.4         3.9          1.3         0.4 setosa    
#>  18          5.1         3.5          1.4         0.3 setosa    
#>  19          5.7         3.8          1.7         0.3 setosa    
#>  20          5.1         3.8          1.5         0.3 setosa    
#>  21          5.4         3.4          1.7         0.2 setosa    
#>  22          5.1         3.7          1.5         0.4 setosa    
#>  23          4.6         3.6          1           0.2 setosa    
#>  24          5.1         3.3          1.7         0.5 setosa    
#>  25          4.8         3.4          1.9         0.2 setosa    
#>  26          5           3            1.6         0.2 setosa    
#>  27          5           3.4          1.6         0.4 setosa    
#>  28          5.2         3.5          1.5         0.2 setosa    
#>  29          5.2         3.4          1.4         0.2 setosa    
#>  30          4.7         3.2          1.6         0.2 setosa    
#>  31          4.8         3.1          1.6         0.2 setosa    
#>  32          5.4         3.4          1.5         0.4 setosa    
#>  33          5.2         4.1          1.5         0.1 setosa    
#>  34          5.5         4.2          1.4         0.2 setosa    
#>  35          4.9         3.1          1.5         0.2 setosa    
#>  36          5           3.2          1.2         0.2 setosa    
#>  37          5.5         3.5          1.3         0.2 setosa    
#>  38          4.9         3.6          1.4         0.1 setosa    
#>  39          4.4         3            1.3         0.2 setosa    
#>  40          5.1         3.4          1.5         0.2 setosa    
#>  41          5           3.5          1.3         0.3 setosa    
#>  42          4.5         2.3          1.3         0.3 setosa    
#>  43          4.4         3.2          1.3         0.2 setosa    
#>  44          5           3.5          1.6         0.6 setosa    
#>  45          5.1         3.8          1.9         0.4 setosa    
#>  46          4.8         3            1.4         0.3 setosa    
#>  47          5.1         3.8          1.6         0.2 setosa    
#>  48          4.6         3.2          1.4         0.2 setosa    
#>  49          5.3         3.7          1.5         0.2 setosa    
#>  50          5           3.3          1.4         0.2 setosa    
#>  51          7           3.2          4.7         1.4 versicolor
#>  52          6.4         3.2          4.5         1.5 versicolor
#>  53          6.9         3.1          4.9         1.5 versicolor
#>  54          5.5         2.3          4           1.3 versicolor
#>  55          6.5         2.8          4.6         1.5 versicolor
#>  56          5.7         2.8          4.5         1.3 versicolor
#>  57          6.3         3.3          4.7         1.6 versicolor
#>  58          4.9         2.4          3.3         1   versicolor
#>  59          6.6         2.9          4.6         1.3 versicolor
#>  60          5.2         2.7          3.9         1.4 versicolor
#>  61          5           2            3.5         1   versicolor
#>  62          5.9         3            4.2         1.5 versicolor
#>  63          6           2.2          4           1   versicolor
#>  64          6.1         2.9          4.7         1.4 versicolor
#>  65          5.6         2.9          3.6         1.3 versicolor
#>  66          6.7         3.1          4.4         1.4 versicolor
#>  67          5.6         3            4.5         1.5 versicolor
#>  68          5.8         2.7          4.1         1   versicolor
#>  69          6.2         2.2          4.5         1.5 versicolor
#>  70          5.6         2.5          3.9         1.1 versicolor
#>  71          5.9         3.2          4.8         1.8 versicolor
#>  72          6.1         2.8          4           1.3 versicolor
#>  73          6.3         2.5          4.9         1.5 versicolor
#>  74          6.1         2.8          4.7         1.2 versicolor
#>  75          6.4         2.9          4.3         1.3 versicolor
#>  76          6.6         3            4.4         1.4 versicolor
#>  77          6.8         2.8          4.8         1.4 versicolor
#>  78          6.7         3            5           1.7 versicolor
#>  79          6           2.9          4.5         1.5 versicolor
#>  80          5.7         2.6          3.5         1   versicolor
#>  81          5.5         2.4          3.8         1.1 versicolor
#>  82          5.5         2.4          3.7         1   versicolor
#>  83          5.8         2.7          3.9         1.2 versicolor
#>  84          6           2.7          5.1         1.6 versicolor
#>  85          5.4         3            4.5         1.5 versicolor
#>  86          6           3.4          4.5         1.6 versicolor
#>  87          6.7         3.1          4.7         1.5 versicolor
#>  88          6.3         2.3          4.4         1.3 versicolor
#>  89          5.6         3            4.1         1.3 versicolor
#>  90          5.5         2.5          4           1.3 versicolor
#>  91          5.5         2.6          4.4         1.2 versicolor
#>  92          6.1         3            4.6         1.4 versicolor
#>  93          5.8         2.6          4           1.2 versicolor
#>  94          5           2.3          3.3         1   versicolor
#>  95          5.6         2.7          4.2         1.3 versicolor
#>  96          5.7         3            4.2         1.2 versicolor
#>  97          5.7         2.9          4.2         1.3 versicolor
#>  98          6.2         2.9          4.3         1.3 versicolor
#>  99          5.1         2.5          3           1.1 versicolor
#> 100          5.7         2.8          4.1         1.3 versicolor
#> # ... with 50 more rows

print(fgeo.data::luquillo_tree5_random)

#> # A tibble: 1,000 x 19
#>    treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID
#>     <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int>
#>  1    104    143 10009 10009   DACE~ 113      10.3  245.    439947        5
#>  2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    466597        5
#>  3    180    225 1001~ 100174  CASA~ 921     165.   410.    466623        5
#>  4    602    736 1006~ 100649  GUAG~ 821     149.   414.    466727        5
#>  5    631    775 10069 10069   PREM~ 213      38.3  245.    439989        5
#>  6    647    793 1007~ 100708  SCHM~ 821     143.   411.    466743        5
#>  7   1086   1339 10122 10122   DRYG~ 413      68.9  253.    440021        5
#>  8   1144   1410 1012~ 101285  SCHM~ 920     161.   395.    466889        5
#>  9   1168   1438 10131 10131   DACE~ 413      70.6  252.    440031        5
#> 10   1380 114352 1015~ 149529  CASA~ 820     142.   386.    466957        5
#> # ... with 990 more rows, and 9 more variables: dbh <dbl>, pom <chr>,
#> #   hom <dbl>, ExactDate <date>, DFstatus <chr>, codes <chr>, nostems <dbl>,
#> #   status <chr>, date <dbl>

print(fgeo.data::luquillo_tree5_random, n_extra = 2)

#> # A tibble: 1,000 x 19
#>    treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID
#>     <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int>
#>  1    104    143 10009 10009   DACE~ 113      10.3  245.    439947        5
#>  2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    466597        5
#>  3    180    225 1001~ 100174  CASA~ 921     165.   410.    466623        5
#>  4    602    736 1006~ 100649  GUAG~ 821     149.   414.    466727        5
#>  5    631    775 10069 10069   PREM~ 213      38.3  245.    439989        5
#>  6    647    793 1007~ 100708  SCHM~ 821     143.   411.    466743        5
#>  7   1086   1339 10122 10122   DRYG~ 413      68.9  253.    440021        5
#>  8   1144   1410 1012~ 101285  SCHM~ 920     161.   395.    466889        5
#>  9   1168   1438 10131 10131   DACE~ 413      70.6  252.    440031        5
#> 10   1380 114352 1015~ 149529  CASA~ 820     142.   386.    466957        5
#> # ... with 990 more rows, and 9 more variables: dbh <dbl>, pom <chr>, ...

print(fgeo.data::luquillo_tree5_random, width = Inf)

#> # A tibble: 1,000 x 19
#>    treeID stemID tag    StemTag sp     quadrat    gx    gy MeasureID CensusID
#>     <int>  <int> <chr>  <chr>   <chr>  <chr>   <dbl> <dbl>     <int>    <int>
#>  1    104    143 10009  10009   DACEXC 113      10.3  245.    439947        5
#>  2    119    158 100104 100104  MYRSPL 1021    183.   410.    466597        5
#>  3    180    225 100171 100174  CASARB 921     165.   410.    466623        5
#>  4    602    736 100649 100649  GUAGUI 821     149.   414.    466727        5
#>  5    631    775 10069  10069   PREMON 213      38.3  245.    439989        5
#>  6    647    793 100708 100708  SCHMOR 821     143.   411.    466743        5
#>  7   1086   1339 10122  10122   DRYGLA 413      68.9  253.    440021        5
#>  8   1144   1410 101285 101285  SCHMOR 920     161.   395.    466889        5
#>  9   1168   1438 10131  10131   DACEXC 413      70.6  252.    440031        5
#> 10   1380 114352 101560 149529  CASARB 820     142.   386.    466957        5
#>      dbh pom     hom ExactDate  DFstatus     codes       nostems status  date
#>    <dbl> <chr> <dbl> <date>     <chr>        <chr>         <dbl> <chr>  <dbl>
#>  1 193   1.4    1.4  2011-11-02 alive        MAIN;A            1 A      18933
#>  2  40.4 1.25   1.26 2012-02-06 alive        MAIN;A            1 A      19029
#>  3  45   1.3    1.32 2012-02-03 alive        SPROUT;A          2 A      19026
#>  4  33   1.3    1.3  2012-02-01 alive        MAIN;A            1 A      19024
#>  5 140   1.25   1.24 2011-11-02 alive        MAIN;A            1 A      18933
#>  6 246   1.25   1.25 2012-01-31 alive        MAIN;A            1 A      19023
#>  7 176   1.45   1.45 2011-11-03 alive        MAIN;A            1 A      18934
#>  8  74   1.3    1.3  2012-02-07 alive        MAIN;A            1 A      19030
#>  9 604   1.3    1.31 2011-11-03 alive        MAIN;A            1 A      18934
#> 10  NA   <NA>  NA    2012-02-06 broken below SPROUT;LS;B       2 A      19029
#> # ... with 990 more rows