Import multiple .rdata files from a directory into a dataframe.

This is shortcut for list_df(rdata_list(x)) with the additional ability to read exclusively .rdata files containing a specific string. It is designed to expect that all .rdata files contain dataframes with the same names -- as is the case for multiple ForestGEO censuses delivered by the database. Notice that function name have the format input_output, i.e. rdata_df.

rdata_df(path_dir, .match = NULL, .id = NULL)

Arguments

path_dir
String; the path to a directory containing the files to read (all must be of appropriate format; see examples).

.match
If not NULL a string to match specific .rdata file-names (excluding the extension).

.id
If not NULL a variable with this name will be created to inform the source file.

Value

A dataframe.

See also

Examples

path_dir <- tool_example("rdata")
dir(path_dir)

#> [1] "tree5.RData" "tree6.RData"

rdata_df(path_dir)

#> # A tibble: 6 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1    104    143 10009 10009   DACE~ 113      10.3  245.    439947        5 193  
#> 2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    466597        5  40.4
#> 3    180    225 1001~ 100174  CASA~ 921     165.   410.    466623        5  45  
#> 4    104    143 10009 10009   DACE~ 113      10.3  245.    582850        6 195  
#> 5    119    158 1001~ 100104  MYRS~ 1021    183.   410.    578696        6  44.9
#> 6    180    225 1001~ 100174  CASA~ 921     165.   410.    617049        6  46.1
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>

rdata_df(path_dir, .match = "6")

#> # A tibble: 3 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1    104    143 10009 10009   DACE~ 113      10.3  245.    582850        6 195  
#> 2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    578696        6  44.9
#> 3    180    225 1001~ 100174  CASA~ 921     165.   410.    617049        6  46.1
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>

dfm <- rdata_df(path_dir, .match = "tree5|6", .id = "source")
dfm

#> # A tibble: 6 x 20
#>   source treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID
#>   <chr>   <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>
#> 1 tree5     104    143 10009 10009   DACE~ 113      10.3  245.    439947
#> 2 tree5     119    158 1001~ 100104  MYRS~ 1021    183.   410.    466597
#> 3 tree5     180    225 1001~ 100174  CASA~ 921     165.   410.    466623
#> 4 tree6     104    143 10009 10009   DACE~ 113      10.3  245.    582850
#> 5 tree6     119    158 1001~ 100104  MYRS~ 1021    183.   410.    578696
#> 6 tree6     180    225 1001~ 100174  CASA~ 921     165.   410.    617049
#> # ... with 10 more variables: CensusID <int>, dbh <dbl>, pom <chr>, hom <dbl>,
#> #   ExactDate <date>, DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>,
#> #   date <dbl>

tail(dfm)

#> # A tibble: 6 x 20
#>   source treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID
#>   <chr>   <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>
#> 1 tree5     104    143 10009 10009   DACE~ 113      10.3  245.    439947
#> 2 tree5     119    158 1001~ 100104  MYRS~ 1021    183.   410.    466597
#> 3 tree5     180    225 1001~ 100174  CASA~ 921     165.   410.    466623
#> 4 tree6     104    143 10009 10009   DACE~ 113      10.3  245.    582850
#> 5 tree6     119    158 1001~ 100104  MYRS~ 1021    183.   410.    578696
#> 6 tree6     180    225 1001~ 100174  CASA~ 921     165.   410.    617049
#> # ... with 10 more variables: CensusID <int>, dbh <dbl>, pom <chr>, hom <dbl>,
#> #   ExactDate <date>, DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>,
#> #   date <dbl>