Read ForestGEO censuses.

This function reads any number of ForestGEO censuses stored in .rdata files and stores them in a nested dataframe of class 'censuses_df', which has a convenient print method and enables using specific methods for working with ForestGEO data.

read_censuses(path_dir, .match = NULL, .id = "censuses")

Arguments

path_dir
String; the path to a directory containing the files to read (all must be of appropriate format; see examples).

.match
If not NULL a string to match specific .rdata file-names (excluding the extension).

.id
If not NULL a variable with this name will be created to inform the source file.

Value

An object of class 'censuses_df'.

See also

Examples

library(dplyr)

censuses <- read_censuses(tool_example("rdata"))
class(censuses)

#> [1] "censuses_df" "tbl_df"      "tbl"         "data.frame" 

censuses$data

#> [[1]]
#> # A tibble: 3 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1    104    143 10009 10009   DACE~ 113      10.3  245.    439947        5 193  
#> 2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    466597        5  40.4
#> 3    180    225 1001~ 100174  CASA~ 921     165.   410.    466623        5  45  
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#> 
#> [[2]]
#> # A tibble: 3 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1    104    143 10009 10009   DACE~ 113      10.3  245.    582850        6 195  
#> 2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    578696        6  44.9
#> 3    180    225 1001~ 100174  CASA~ 921     165.   410.    617049        6  46.1
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#> 

censuses %>%
  pick(dbh > 600) %>%
  pull(data)

#> [[1]]
#> # A tibble: 0 x 19
#> # ... with 19 variables: treeID <int>, stemID <int>, tag <chr>, StemTag <chr>,
#> #   sp <chr>, quadrat <chr>, gx <dbl>, gy <dbl>, MeasureID <int>,
#> #   CensusID <int>, dbh <dbl>, pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#> 
#> [[2]]
#> # A tibble: 0 x 19
#> # ... with 19 variables: treeID <int>, stemID <int>, tag <chr>, StemTag <chr>,
#> #   sp <chr>, quadrat <chr>, gx <dbl>, gy <dbl>, MeasureID <int>,
#> #   CensusID <int>, dbh <dbl>, pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#>