Read ViewFullTable and ViewTaxonomy files delivered by ForestGEO's database.

These functions pre-fill arguments of readr::read_delim() to avoid the most common causes of problems when reading ForestGEO data as delivered by the database. In most cases you should only need to provide the argument file.

read_vft(file, delim = "\t", na = c("", "NA", "NULL"), ...)

read_taxa(file, delim = "\t", na = c("", "NA", "NULL"), ...)

Arguments

file
Either a path to a file, a connection, or literal data (either a single string or a raw vector).

Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded and decompressed.

Literal data is most useful for examples and tests. It must contain at least one new line to be recognised as data (instead of a path).

delim
Single character used to separate fields within a record.

na
Character vector of strings to use for missing values. Set this option to character() to indicate no missing values.

...
Other arguments passed to readr::read_delim().

Value

A dataframe (or tibble).

See also

Examples

read_vft("http://bit.ly/fgeo-data-luquillo-vft-random")

#> # A tibble: 5,921 x 32
#>     DBHID PlotName PlotID Family      Genus     SpeciesName Mnemonic Subspecies
#>     <int> <chr>     <int> <chr>       <chr>     <chr>       <chr>    <chr>     
#>  1    143 luquillo      1 Burseraceae Dacryodes excelsa     DACEXC   <NA>      
#>  2 103518 luquillo      1 Burseraceae Dacryodes excelsa     DACEXC   <NA>      
#>  3 217487 luquillo      1 Burseraceae Dacryodes excelsa     DACEXC   <NA>      
#>  4 347333 luquillo      1 Burseraceae Dacryodes excelsa     DACEXC   <NA>      
#>  5 439947 luquillo      1 Burseraceae Dacryodes excelsa     DACEXC   <NA>      
#>  6 582850 luquillo      1 Burseraceae Dacryodes excelsa     DACEXC   <NA>      
#>  7    158 luquillo      1 Myrtaceae   Myrcia    splendens   MYRSPL   <NA>      
#>  8 103533 luquillo      1 Myrtaceae   Myrcia    splendens   MYRSPL   <NA>      
#>  9 217498 luquillo      1 Myrtaceae   Myrcia    splendens   MYRSPL   <NA>      
#> 10 384255 luquillo      1 Myrtaceae   Myrcia    splendens   MYRSPL   <NA>      
#> # ... with 5,911 more rows, and 24 more variables: SpeciesID <int>,
#> #   SubspeciesID <chr>, QuadratName <chr>, QuadratID <int>, PX <dbl>, PY <dbl>,
#> #   QX <dbl>, QY <dbl>, TreeID <int>, Tag <chr>, StemID <int>,
#> #   StemNumber <int>, StemTag <int>, PrimaryStem <chr>, CensusID <int>,
#> #   PlotCensusNumber <int>, DBH <dbl>, HOM <dbl>, ExactDate <date>, Date <int>,
#> #   ListOfTSM <chr>, HighHOM <int>, LargeStem <chr>, Status <chr>

read_vft("http://bit.ly/fgeo-data-luquillo-taxa")

#> Warning: The following named parsers don't match the column names: DBHID, PlotName, PlotID, QuadratName, QuadratID, PX, PY, QX, QY, TreeID, Tag, StemID, StemNumber, StemTag, PrimaryStem, CensusID, PlotCensusNumber, DBH, HOM, ExactDate, Date, ListOfTSM, HighHOM, LargeStem, Status

#> # A tibble: 163 x 21
#>    ViewID SpeciesID SubspeciesID Family      Mnemonic Genus   SpeciesName Rank 
#>     <int>     <int> <chr>        <chr>       <chr>    <chr>   <chr>       <chr>
#>  1      1        56 <NA>         Fabaceae    AESAME   Aeschy~ americana   <NA> 
#>  2      2        57 <NA>         Euphorbiac~ ALCFLO   Alchor~ floribunda  <NA> 
#>  3      3        58 <NA>         Euphorbiac~ ALCLAT   Alchor~ latifolia   <NA> 
#>  4      4        59 <NA>         Fabaceae-p~ ANDINE   Andira  inermis     <NA> 
#>  5      5        60 <NA>         Rubiaceae   ANTOBT   Stenos~ obtusifoli~ <NA> 
#>  6      6        61 <NA>         Myrsinaceae ARDGLA   Ardisia glauciflora <NA> 
#>  7      7        62 <NA>         Moraceae    ARTALT   Artoca~ altilis     <NA> 
#>  8      8        63 <NA>         Lauraceae   BEIPEN   Beilsc~ pendula     <NA> 
#>  9      9        64 <NA>         Solanaceae  BRUPOR   Brunfe~ portoricen~ <NA> 
#> 10     10        65 <NA>         Combretace~ BUCTET   Buchen~ tetraphylla <NA> 
#> # ... with 153 more rows, and 13 more variables: Subspecies <chr>,
#> #   Authority <chr>, IDLevel <chr>, subspMnemonic <chr>, subspAuthority <chr>,
#> #   FieldFamily <chr>, Lifeform <chr>, Description <chr>, wsg <dbl>,
#> #   wsglevel <chr>, ListOfOldNames <chr>, Specimens <chr>, Reference <chr>