If necessary, pull elevation data and rename x and y to gx and gy.

This functions helps standarize elevation data. Elevation data may be stored as a dataframe in the element col of a list or may be directly the dataframe. Also, the names may be x and y wich is inconsistent with the names gx and gy of census datasets. Whatever the structure of the input (either a dataframe or a list) and whatever the names (x and y, or gx and gy), this function outputs a dataframe with names gx and gy, and checks that the variable elev is present.

restructure_elev(elevation)

Arguments

elevation
A dataframe or list of ForestGEO's elevation data.

Value

A dataframe

Acknowledgements

This function was inspired by David Kenfack.

Examples

# Dataframe input
elev_df <- bciex::bci_elevation
str(elev_df)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	20301 obs. of  3 variables:
#>  $ x   : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ y   : int  0 5 10 15 20 25 30 35 40 45 ...
#>  $ elev: num  121 121 121 121 121 ...

elev_clean <- restructure_elev(elev_df)
str(elev_clean)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	20301 obs. of  3 variables:
#>  $ gx  : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ gy  : int  0 5 10 15 20 25 30 35 40 45 ...
#>  $ elev: num  121 121 121 121 121 ...

# List input
elev_list <- list(col = elev_df, other = "stuff")
str(elev_list)

#> List of 2
#>  $ col  :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	20301 obs. of  3 variables:
#>   ..$ x   : int [1:20301] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ y   : int [1:20301] 0 5 10 15 20 25 30 35 40 45 ...
#>   ..$ elev: num [1:20301] 121 121 121 121 121 ...
#>  $ other: chr "stuff"

elev_clean <- restructure_elev(elev_list)
str(elev_clean)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	20301 obs. of  3 variables:
#>  $ gx  : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ gy  : int  0 5 10 15 20 25 30 35 40 45 ...
#>  $ elev: num  121 121 121 121 121 ...