Sanitize ViewFullTable and ViewTaxonomy.

These functions sanitize dataframes from ViewFullTable and ViewTaxonomy to achieve the same effect as reading with read_vft() and read_taxa(). The most common problems that these functions fix are these:

Ensure that each column has the correct type.

Ensure that missing values are represented with NAs -- not with the literal string "NULL".

sanitize_vft(x, col_types = type_vft(), na = c("", "NA", "NULL"), ...)

sanitize_taxa(x, col_types = type_taxa(), na = c("", "NA", "NULL"), ...)

Arguments

x
A dataframe; either a ForestGEO ViewFullTable (sanitize_vft()) or ViewTaxonomy (sanitize_vft()).

col_types
One of NULL, a cols() specification, or a string. See vignette("column-types") for more details.

If NULL, all column types will be imputed from the first 1000 rows on the input. This is convenient (and fast), but not robust. If the imputation fails, you'll need to supply the correct types yourself.

If a column specification created by cols(), it must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Unlike other functions type_convert() does not allow character specifications of col_types.

na
Character vector of strings to use for missing values. Set this option to character() to indicate no missing values.

...
Arguments passed on to readr::type_convert

df

A data frame.

col_types

One of NULL, a cols() specification, or a string. See vignette("column-types") for more details. If NULL, all column types will be imputed from the first 1000 rows on the input. This is convenient (and fast), but not robust. If the imputation fails, you'll need to supply the correct types yourself. If a column specification created by cols(), it must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only(). Unlike other functions type_convert() does not allow character specifications of col_types.

na

Character vector of strings to use for missing values. Set this option to character() to indicate no missing values.

trim_ws

Should leading and trailing whitespace be trimmed from each field before parsing it?

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

Value

A dataframe.

Acknowledgments

Thanks to Shameema Jafferjee Esufali for motivating this functions.

See also

Examples

vft <- fgeo.data::luquillo_vft_4quad

# Introdue problems to show how to fix them
# Bad column types
vft[] <- lapply(vft, as.character)
# Bad representation of missing values
vft$PlotName <- "NULL"

# "NULL" should be replaced by `NA` and `DBH` should be numeric
str(vft[c("PlotName", "DBH")])

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	3302 obs. of  2 variables:
#>  $ PlotName: chr  "NULL" "NULL" "NULL" "NULL" ...
#>  $ DBH     : chr  NA "185" NA "341" ...

# Fix
vft_sane <- sanitize_vft(vft)
str(vft_sane[c("PlotName", "DBH")])

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	3302 obs. of  2 variables:
#>  $ PlotName: chr  NA NA NA NA ...
#>  $ DBH     : num  NA 185 NA 341 370 375 167 172 187 141 ...

taxa <- fgeo.data::luquillo_taxa
# Bad column types
taxa[] <- lapply(taxa, as.character)
# Bad representation of missing values
taxa$SubspeciesID <- "NULL"

# "NULL" should be replaced by `NA` and `ViewID` should be integer
str(taxa[c("SubspeciesID", "ViewID")])

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	163 obs. of  2 variables:
#>  $ SubspeciesID: chr  "NULL" "NULL" "NULL" "NULL" ...
#>  $ ViewID      : chr  "1" "2" "3" "4" ...

# Fix
taxa_sane <- sanitize_taxa(taxa)
str(taxa_sane[c("SubspeciesID", "ViewID")])

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	163 obs. of  2 variables:
#>  $ SubspeciesID: chr  NA NA NA NA ...
#>  $ ViewID      : int  1 2 3 4 5 6 7 8 9 10 ...