Sanitize ViewFullTable and ViewTaxonomy.

These functions sanitize dataframes from ViewFullTable and ViewTaxonomy to achieve the same effect as reading with read_vft() and read_taxa(). The most common problems that these functions fix are these:

Ensure that each column has the correct type.

Ensure that missing values are represented with NAs -- not with the literal string "NULL".

sanitze_vft(vft)

Arguments

vft
A dataframe; particularly a ForestGEO ViewFullTable (vft).

taxa
A dataframe; particularly a ForestGEO ViewTaxonomy (taxa).

Acknowledgments

Thanks to Shameema Jafferjee Esufali for motivating this functions.

See also

Examples

# Example data
vft <- tibble::tibble(QuadratName = c(0000, 0001), DBH = c("NULL", "NULL"))
str(vft)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	2 obs. of  2 variables:
#>  $ QuadratName: num  0 1
#>  $ DBH        : chr  "NULL" "NULL"

# Warns because lot's of columns are missing
sane <- sanitze_vft(vft)

#> Warning: The following named parsers don't match the column names: DBHID, PlotName, PlotID, Family, Genus, SpeciesName, Mnemonic, Subspecies, SpeciesID, SubspeciesID, QuadratID, PX, PY, QX, QY, TreeID, Tag, StemID, StemNumber, StemTag, PrimaryStem, CensusID, PlotCensusNumber, HOM, ExactDate, Date, ListOfTSM, HighHOM, LargeStem, Status

str(sane, give.attr = FALSE)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	2 obs. of  2 variables:
#>  $ QuadratName: chr  "0" "1"
#>  $ DBH        : num  NA NA

taxa <- luquillo::ViewTaxonomy_luquillo
# Introduce mistakes
taxa$SubspeciesID <- "NULL"
taxa$SpeciesID <- as.double(taxa$SpeciesID)
str(taxa[1:3], give.attr = FALSE)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	163 obs. of  3 variables:
#>  $ ViewID      : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ SpeciesID   : num  56 57 58 59 60 61 62 63 64 65 ...
#>  $ SubspeciesID: chr  "NULL" "NULL" "NULL" "NULL" ...

sane <- sanitze_taxa(taxa)
str(sane[1:3], give.attr = FALSE)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	163 obs. of  3 variables:
#>  $ ViewID      : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ SpeciesID   : int  56 57 58 59 60 61 62 63 64 65 ...
#>  $ SubspeciesID: chr  NA NA NA NA ...