Suffix a strings where a vector exactly matches one specific string.

Suffix a strings where a vector exactly matches one specific string.

str_suffix_match(string, to_match, .match, suffix)

Arguments

string
A vector of strings to suffix.

to_match
A vector of strings to match with .match.

.match
A string to match the vector to_match with.

suffix
A suffix to add at the end of each element of string.

Value

A modified version of x.

Examples

str_suffix_match(
  string = c("tag1", "tag2"),
  to_match = c("dead", "not-dead"),
  .match = "dead",
  suffix = ".d"
)

#> [1] "tag1.d" "tag2"  

library(dplyr)
vft <- tribble(
  ~Tag, ~Status,
  "01", "dead",
  "02", "alive"
)
mutate(vft, tagged = str_suffix_match(Tag, Status, "dead", ".d"))

#> # A tibble: 2 x 3
#>   Tag   Status tagged
#>   <chr> <chr>  <chr> 
#> 1 01    dead   01.d  
#> 2 02    alive  02