Help to read ForestGEO data safely, with consistent column type.

A common cause of problems is feeding functions with data which columns are not all of the expected type. The problem often begins when reading data from a text file with functions such as utils::read.csv(), utils::read.delim(), and friends -- which commonly guess wrongly the column type that you more likely expect. These common offenders are strongly discouraged; instead consider using readr::read_csv(), readr::read_csv(), and friends, which guess column types correctly much more often than their analogs from the utils package.

The functions type_vft(), type_taxa(), and friends help you to read data more safely by explicitly specifying what type to expect from each column of known datasets. These functions output the specification for the argument col_types() of the functions readr::read_*():

type_vft(): Type specification for ViewFullTable.

type_taxa(): Type specification for ViewFullTaxonomy.

type_vft()

type_taxa()

Value

A list.

Details

Types reference (for more details see read_delim()):

c = character,

i = integer,

n = number,

d = double,

l = logical,

D = date,

T = date time,

t = time,

? = guess,

or _/- to skip the column.'.

See also

Examples

str(type_vft())

#> List of 32
#>  $ DBHID           : chr "i"
#>  $ PlotName        : chr "c"
#>  $ PlotID          : chr "i"
#>  $ Family          : chr "c"
#>  $ Genus           : chr "c"
#>  $ SpeciesName     : chr "c"
#>  $ Mnemonic        : chr "c"
#>  $ Subspecies      : chr "c"
#>  $ SpeciesID       : chr "i"
#>  $ SubspeciesID    : chr "c"
#>  $ QuadratName     : chr "c"
#>  $ QuadratID       : chr "i"
#>  $ PX              : chr "d"
#>  $ PY              : chr "d"
#>  $ QX              : chr "d"
#>  $ QY              : chr "d"
#>  $ TreeID          : chr "i"
#>  $ Tag             : chr "c"
#>  $ StemID          : chr "i"
#>  $ StemNumber      : chr "i"
#>  $ StemTag         : chr "i"
#>  $ PrimaryStem     : chr "c"
#>  $ CensusID        : chr "i"
#>  $ PlotCensusNumber: chr "i"
#>  $ DBH             : chr "d"
#>  $ HOM             : chr "d"
#>  $ ExactDate       : chr "D"
#>  $ Date            : chr "i"
#>  $ ListOfTSM       : chr "c"
#>  $ HighHOM         : chr "i"
#>  $ LargeStem       : chr "c"
#>  $ Status          : chr "c"

# NOT RUN {
# Run this with your own ViewFullTable
readr::read_csv("ViewFullTable.csv", col_types = type_vft())
# }

str(type_taxa())

#> List of 21
#>  $ ViewID        : chr "i"
#>  $ SpeciesID     : chr "i"
#>  $ SubspeciesID  : chr "c"
#>  $ Family        : chr "c"
#>  $ Mnemonic      : chr "c"
#>  $ Genus         : chr "c"
#>  $ SpeciesName   : chr "c"
#>  $ Rank          : chr "c"
#>  $ Subspecies    : chr "c"
#>  $ Authority     : chr "c"
#>  $ IDLevel       : chr "c"
#>  $ subspMnemonic : chr "c"
#>  $ subspAuthority: chr "c"
#>  $ FieldFamily   : chr "c"
#>  $ Lifeform      : chr "c"
#>  $ Description   : chr "c"
#>  $ wsg           : chr "d"
#>  $ wsglevel      : chr "c"
#>  $ ListOfOldNames: chr "c"
#>  $ Specimens     : chr "c"
#>  $ Reference     : chr "c"

# NOT RUN {
# Run this with your own ViewTaxonomy
readr::read_delim("ViewTaxonomy.csv", col_types = type_taxa())
# }