Combine spreadsheets from excel workbooks and output common data structures.

These functions combine spreadsheets from excel workbooks into common data structures. xl_sheets_to_csv() and xl_sheets_to_xl() write a .csv or excel (.xlsx) file per workbook -- combining all spreadsheets. xl_sheets_to_df outputs a list where each dataframes combines all spreadsheeets of a workbook.

xl_sheets_to_csv(input_dir, output_dir = "./", first_census = FALSE)

xl_sheets_to_xl(input_dir, output_dir = "./", first_census = FALSE)

xl_sheets_to_df(input_dir, first_census = FALSE)

Arguments

input_dir
String giving the directory containing the excel workbooks to read from.

output_dir
String giving the directory where to write .csv files to.

first_census
This argument tells these functions what sheets to expect in the input.

Use TRUE if this is your first census. The expected input must have sheets (1) "root", (2) "multi_stems", (3) "secondary_stems", and (4) "single_stems".

Use FALSE (default) if this is not your first census. The expected input must have sheets (1) "root", (2) "original_stems", (3) "new_secondary_stems", and (4) "recruits".

Value

Writes one .csv file for each workbook.

Details

This is a rigid function with a very specific goal: To process data from a specific sampling software -- FastField. Specifically, this is what this function does:

Reads each spreadsheet from each workbook and map it to a dataframe.

Lowercases and links the names of each dataframe.

Keeps only these dataframes: (1) original_stems, (2) new_secondary_stems, and (3) "recruits".

Dates the data by submission_id (date comes from the spreadsheet root).

Lowercases and links the names of each dataframe-variable.

Drops fake stems.

Output a common data structure of your choice.

Acknowledgment

Sabrina Russo helped to make these functions useful with first censuses.

Examples

library(fs)
library(readr)
library(readxl)

# NOT A FIRST CENSUS
# Path to the folder I want to read excel files from
input_dir <- dirname(example_path("two_files/new_stem_1.xlsx"))
input_dir

#> [1] "C:/Users/LeporeM/Dropbox/git_repos/fgeo.tool/inst/extdata/two_files"

# Files I want to read
dir(input_dir, pattern = "xlsx")

#> [1] "new_stem_1.xlsx" "new_stem_2.xlsx"

# Path to the folder I want to write .csv files to
output_dir <- tempdir()

# Output a csv file
xl_sheets_to_csv(input_dir, output_dir)

# Confirm
path_file(dir_ls(output_dir, regexp = "new_stem.*csv$"))

#> new_stem_1.csv new_stem_2.csv 

# Also possible to output excel and a list of dataframe. See next section.

# FIRST CENSUS
input_dir <- dirname(example_path("first_census/census.xlsx"))
# As a reminder you'll get a warning of missing sheets
# Output list of dataframes (one per input workbook -- here only one)
xl_sheets_to_df(input_dir, first_census = TRUE)

#> Warning: `.x` has no dataframe new_secondary_stems

#> Warning: `.x` has no dataframe recruits

#> $census.xlsx
#> # A tibble: 49 x 15
#>    submission_id quadrat tag   stem_tag species species_code dbh   status codes
#>    <chr>         <chr>   <chr> <chr>    <chr>   <chr>        <chr> <chr>  <chr>
#>  1 c01b37d5-2c3~ 0816    007   1        Americ~ TILIAM       51.2  LI     M    
#>  2 efd8e991-2e3~ 0916    004   1        Americ~ TILIAM       31.6  LI     M    
#>  3 efd8e991-2e3~ 0916    007   1        Ironwo~ OSTRVI       10.2  LI     M    
#>  4 efd8e991-2e3~ 0916    008   1        Ironwo~ OSTRVI       2.7   LI     M;L  
#>  5 efd8e991-2e3~ 0916    010   1        Hackbe~ CELTOC       25.9  LI     M    
#>  6 efd8e991-2e3~ 0916    012   1        Bur Oak QUERMA       35.2  LI     M;FS~
#>  7 7997596a-c94~ 1017    001   1        Americ~ TILIAM       65.3  LI     M;FS~
#>  8 572ddae1-d02~ 0816    002   1        Hackbe~ CELTOC       45.1  LI     M    
#>  9 c01b37d5-2c3~ 0816    007   2        TILIAM  <NA>         42.4  LI     S;L  
#> 10 c01b37d5-2c3~ 0816    007   3        TILIAM  <NA>         39.8  LI     S    
#> # ... with 39 more rows, and 6 more variables: notes <chr>, pom <chr>,
#> #   sheet <chr>, section_id <chr>, unique_stem <chr>, date <chr>
#> 

# Output excel
xl_sheets_to_xl(input_dir, output_dir, first_census = TRUE)

#> Warning: `.x` has no dataframe new_secondary_stems

#> Warning: `.x` has no dataframe recruits

# Read back
filename <- path(output_dir, "census.xlsx")
out <- read_excel(filename)
str(out, give.attr = FALSE)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	49 obs. of  15 variables:
#>  $ submission_id: chr  "c01b37d5-2c30-4086-b4a1-78327dfa0a9a" "efd8e991-2e34-40c8-b503-44def6e542d9" "efd8e991-2e34-40c8-b503-44def6e542d9" "efd8e991-2e34-40c8-b503-44def6e542d9" ...
#>  $ quadrat      : chr  "0816" "0916" "0916" "0916" ...
#>  $ tag          : chr  "007" "004" "007" "008" ...
#>  $ stem_tag     : chr  "1" "1" "1" "1" ...
#>  $ species      : chr  "American Basswood/Linden" "American Basswood/Linden" "Ironwood" "Ironwood" ...
#>  $ species_code : chr  "TILIAM" "TILIAM" "OSTRVI" "OSTRVI" ...
#>  $ dbh          : chr  "51.2" "31.6" "10.2" "2.7" ...
#>  $ status       : chr  "LI" "LI" "LI" "LI" ...
#>  $ codes        : chr  "M" "M" "M" "M;L" ...
#>  $ notes        : chr  NA NA NA "In brush pile" ...
#>  $ pom          : chr  "1.3" "1.3" "1.3" "1.3" ...
#>  $ sheet        : chr  "multi_stems" "multi_stems" "multi_stems" "multi_stems" ...
#>  $ section_id   : chr  NA NA NA NA ...
#>  $ unique_stem  : chr  "007_1" "004_1" "007_1" "008_1" ...
#>  $ date         : chr  "2018.05.14" "2018.05.14" "2018.05.14" "2018.05.14" ...