Add columns lx/ly, QX/QY, index, col/row, hectindex, quad, gx/gy.

These functions add columns to position trees in a forest plot. They work with ViewFullTable, tree and stem tables. From the input table, most functions use only the gx and gy columns (or equivalent columns). The exception is the function add_gxgy() which inputs quadrat information. If your data lacks some important column, an error message will inform you which column is missing.

add_lxly(x, gridsize = 20, plotdim = NULL)

add_qxqy(x, gridsize = 20, plotdim = NULL)

add_index(x, gridsize = 20, plotdim = NULL)

add_col_row(x, gridsize = 20, plotdim = NULL)

add_hectindex(x, gridsize = 20, plotdim = NULL)

add_quad(x, gridsize = 20, plotdim = NULL, start = NULL, width = 2)

add_gxgy(x, gridsize = 20, start = 0)

Arguments

x
A ForestGEO-like dataframe: A ViewFullTable, tree or stem table.

gridsize
The gridsize of the census plot (commonly 20 m).

plotdim
The global dimensions of the census plot (i.e. the maximum possible values of gx and gy).

start
Defaults to label the first quadrat as "0101". Use 0 to label it as "0000" instead.

width
Number; width to pad the labels of plot-columns and -rows.

Value

For any given var, a function add_var() returns a modified version of the input dataframe, with the additional variable(s) var.

Details

These functions are adapted from the CTFS R Package.

See also

Examples

x <- tribble(
    ~gx,    ~gy,
      0,      0,
     50,     25,
  999.9, 499.95,
   1000,    500
)

# `gridsize` has a common default; `plotdim` is guessed from the data
add_lxly(x)

#> Guessing: plotdim = c(1000, 500)

#> * If guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 4
#>      gx    gy    lx    ly
#>   <dbl> <dbl> <dbl> <dbl>
#> 1    0     0    0     0  
#> 2   50    25   10     5  
#> 3 1000.  500.  19.9  19.9
#> 4 1000   500   NA    NA  

gridsize <- 20
plotdim <- c(1000, 500)

add_qxqy(x, gridsize, plotdim)

#> # A tibble: 4 x 4
#>      gx    gy    QX    QY
#>   <dbl> <dbl> <dbl> <dbl>
#> 1    0     0    0     0  
#> 2   50    25   10     5  
#> 3 1000.  500.  19.9  19.9
#> 4 1000   500   NA    NA  

add_index(x, gridsize, plotdim)

#> # A tibble: 4 x 3
#>      gx    gy index
#>   <dbl> <dbl> <dbl>
#> 1    0     0      1
#> 2   50    25     52
#> 3 1000.  500.  1250
#> 4 1000   500     NA

add_hectindex(x, gridsize, plotdim)

#> Warning: Some values of `gx` and/or `gy` lay at or beyond plot limits

#> # A tibble: 4 x 3
#>      gx    gy hectindex
#>   <dbl> <dbl>     <dbl>
#> 1    0     0          1
#> 2   50    25          1
#> 3 1000.  500.        50
#> 4 1000   500         56

add_quad(x, gridsize, plotdim)

#> # A tibble: 4 x 3
#>      gx    gy quad 
#>   <dbl> <dbl> <chr>
#> 1    0     0  0101 
#> 2   50    25  0302 
#> 3 1000.  500. 5025 
#> 4 1000   500  NA   

add_quad(x, gridsize, plotdim, start = 0)

#> # A tibble: 4 x 3
#>      gx    gy quad 
#>   <dbl> <dbl> <chr>
#> 1    0     0  0000 
#> 2   50    25  0201 
#> 3 1000.  500. 4924 
#> 4 1000   500  NA   

# `width` gives the nuber of digits to pad the label of plot-rows and
# plot-columns, e.g. 3 pads plot-rows with three zeros and plot-columns with
# an extra trhree zeros, resulting in a total of 6 zeros.
add_quad(x, gridsize, plotdim, start = 0, width = 3)

#> # A tibble: 4 x 3
#>      gx    gy quad  
#>   <dbl> <dbl> <chr> 
#> 1    0     0  000000
#> 2   50    25  002001
#> 3 1000.  500. 049024
#> 4 1000   500  NA    

add_col_row(x, gridsize, plotdim)

#> # A tibble: 4 x 4
#>      gx    gy col   row  
#>   <dbl> <dbl> <chr> <chr>
#> 1    0     0  01    01   
#> 2   50    25  03    02   
#> 3 1000.  500. 50    25   
#> 4 1000   500  NA    NA   

# From `quadrat` or `QuadratName` --------------------------------------
x <- tribble(
  ~QuadratName,
        "0001",
        "0011",
        "0101",
        "1001"
)

# Output `gx` and `gy` ---------------

add_gxgy(x)

#> # A tibble: 4 x 3
#>   QuadratName    gx    gy
#>   <chr>       <dbl> <dbl>
#> 1 0001            0    20
#> 2 0011            0   220
#> 3 0101           20    20
#> 4 1001          200    20

# Warning: The data may already have `gx` and `gx` columns
gxgy <- add_gxgy(fgeo.x::tree5)
select(gxgy, matches("gx|gy"))

#> # A tibble: 30 x 4
#>       gx    gy   gx1   gy1
#>    <dbl> <dbl> <dbl> <dbl>
#>  1 139.  425.    140   440
#>  2  94.8 424.    100   440
#>  3  61.3 496.     80   500
#>  4 100.  328.    120   340
#>  5  53.8  73.8    60    80
#>  6 203.  110.    220   120
#>  7 172.   14.7   180    20
#>  8 184.  194.    200   200
#>  9 191.  132.    200   140
#> 10 274.  279.    280   280
#> # ... with 20 more rows

# Output `col` and `row` -------------

# Create columns `col` and `row` from `QuadratName` with `tidyr::separate()`
# The argument `sep` lets you separate `QuadratName` at any positon

# NOT RUN {
tidyr_is_installed <- requireNamespace("tidyr", quietly = TRUE)
stringr_is_installed <- requireNamespace("stringr", quietly = TRUE)

if (tidyr_is_installed && stringr_is_installed) {
  library(tidyr)
  library(stringr)

  vft <- tibble(QuadratName = c("0001", "0011"))
  vft

  separate(
    vft,
    QuadratName, into = c("col", "row"),
    sep = 2
  )

  census <- select(fgeo.x::tree5, quadrat)
  census

  census$quadrat <- str_pad(census$quadrat, width = 4, pad = 0)

  separate(
    census,
    quadrat, into = c("col", "row"),
    sep = 2,
    remove = FALSE
  )
}
# }