Add columns to a ForestGEO dataframe to position stems in a forest plot.

These functions mostly wrap legacy code from the CTFS R Package.

add_lxly(x, gridsize = 20, plotdim = NULL)

add_qxqy(x, gridsize = 20, plotdim = NULL)

add_index(x, gridsize = 20, plotdim = NULL)

add_col_row(x, gridsize = 20, plotdim = NULL)

add_hectindex(x, gridsize = 20, plotdim = NULL)

add_quad(x, gridsize = 20, plotdim = NULL, start = 1, width = 2)

add_col_row2(x)

Arguments

x
A ForestGEO table, either a ViewFullTable or a census table.

gridsize
The gridsize of the census plot (commonly 20 m).

plotdim
The global dimensions of the census plot (i.e. the maximum possible values of gx and gy).

start
1 or 0, indicating how to label the first plot-column.

width
Number; width to pad the labels of plot-columns and -rows (passed to stringr::str_pad()).

Value

A modified version of the dataframe x with the additional variable(s) var.

Details

These functions add columns to locate stems in a forest plot. add_col_row() and add_col_row2() differ in that the former inputs plot coordinates and the later inputs QuadratName.

See also

Examples

x <- tibble::tibble(gx = c(0, 50, 999.9, 1000), gy = gx/2)

add_lxly(x)

#> Gessing: plotdim = c(1000, 500)

#>   * If guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 4
#>      gx    gy    lx    ly
#>   <dbl> <dbl> <dbl> <dbl>
#> 1    0     0    0     0  
#> 2   50    25   10     5  
#> 3 1000.  500.  19.9  19.9
#> 4 1000   500   NA    NA  

add_qxqy(x)

#> Gessing: plotdim = c(1000, 500)

#>   * If guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 4
#>      gx    gy    QX    QY
#>   <dbl> <dbl> <dbl> <dbl>
#> 1    0     0    0     0  
#> 2   50    25   10     5  
#> 3 1000.  500.  19.9  19.9
#> 4 1000   500   NA    NA  

add_index(x)

#> Gessing: plotdim = c(1000, 500)

#>   * If guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 3
#>      gx    gy index
#>   <dbl> <dbl> <dbl>
#> 1    0     0      1
#> 2   50    25     52
#> 3 1000.  500.  1250
#> 4 1000   500     NA

add_hectindex(x)

#> Gessing: plotdim = c(1000, 500)

#>   * If guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 3
#>      gx    gy hectindex
#>   <dbl> <dbl>     <dbl>
#> 1    0     0          1
#> 2   50    25          1
#> 3 1000.  500.        50
#> 4 1000   500         56

add_quad(x)

#> Gessing: plotdim = c(1000, 500)

#>   * If guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 3
#>      gx    gy quad 
#>   <dbl> <dbl> <chr>
#> 1    0     0  0101 
#> 2   50    25  0302 
#> 3 1000.  500. 5025 
#> 4 1000   500  NANA 

add_quad(x, start = 0)

#> Gessing: plotdim = c(1000, 500)

#>   * If guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 3
#>      gx    gy quad 
#>   <dbl> <dbl> <chr>
#> 1    0     0  0000 
#> 2   50    25  0201 
#> 3 1000.  500. 4924 
#> 4 1000   500  NANA 

# `width` gives the nuber of digits to pad the label of plot-rows and
# plot-columns, e.g. 3 pads plot-rows with three zeros and plot-columns with an
# extra trhree zeros, resulting in a total of 6 zeros.
add_quad(x, start = 0, width = 3)

#> Gessing: plotdim = c(1000, 500)

#>   * If guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 3
#>      gx    gy quad  
#>   <dbl> <dbl> <chr> 
#> 1    0     0  000000
#> 2   50    25  002001
#> 3 1000.  500. 049024
#> 4 1000   500  NANA  

add_col_row(x)

#> Gessing: plotdim = c(1000, 500)

#>   * If guess is wrong, provide the correct argument `plotdim`

#> # A tibble: 4 x 4
#>      gx    gy col   row  
#>   <dbl> <dbl> <chr> <chr>
#> 1    0     0  01    01   
#> 2   50    25  03    02   
#> 3 1000.  500. 50    25   
#> 4 1000   500  <NA>  <NA> 

# Column and row from QuadratName
x <- tibble::tribble(
  ~QuadratName,
  "0001",
  "0011",
  "0101",
  "1001"
)
add_col_row2(x)

#>   QuadratName  col  row
#> 1        0001 0001 0001
#> 2        0011 0011 0011
#> 3        0101 0101 0101
#> 4        1001 1001 1001