Create field forms forms.
Source: R/fieldforms.R

Create field forms. Data should be from a single plot and of a single census. Pandoc must be installed (see section Installing Pandoc).

fieldforms_output(forms, filename = "fieldforms.docx",
  output_format = "word_document", header = fieldforms_header(),
  output_dir = getwd())

Arguments

forms
A list of dataframes -- most likely the output of fieldforms_prepare().

filename
String; path to the output file. Must include the extension ".docx". Passed to rmarkdown::render().

output_format
String. Currently only "word_document" is supported.

header
A string to use as the header of the fieldform of each quadrat.

output_dir
String giving a directory to save the file in.

Value

A (MS Word) .docx file.

Installing Pandoc

This function requires that you have pandoc installed. See installation instructions at https://pandoc.org/installing.html. To install pandoc from R see the function install.pandoc() of the installr package (https://goo.gl/cpaxtd).

See also

fieldforms_header().

Examples

library(fgeo.tool)
x <- top4quad
prep <- fieldforms_prepare(x)
lapply(prep, head)

#> $`4915`
#> # A tibble: 6 x 12
#>   sq    tag    stemid mnemonic   dbh   hom status .dbh  code  .hom  note  comm 
#>   <chr> <chr>   <int> <chr>    <int> <dbl> <chr>  <chr> <chr> <chr> <chr> <chr>
#> 1 11    124872      4 hybapr      14  1.30 alive  ""    ""    ""    ""    ""   
#> 2 11    127645      1 swars2      28  1.30 alive  ""    ""    ""    ""    ""   
#> 3 11    127646      1 faraoc      57  1.30 alive  ""    ""    ""    ""    ""   
#> 4 11    127647      6 cappfr      NA NA    alive  ""    ""    ""    ""    ""   
#> 5 11    127649     18 hybapr      15  1.30 alive  ""    ""    ""    ""    ""   
#> 6 11    127649     19 hybapr      14  1.30 alive  ""    ""    ""    ""    ""   
#> 
#> $`4916`
#> # A tibble: 6 x 12
#>   sq    tag    stemid mnemonic   dbh   hom status .dbh  code  .hom  note  comm 
#>   <chr> <chr>   <int> <chr>    <int> <dbl> <chr>  <chr> <chr> <chr> <chr> <chr>
#> 1 11    022900      1 posola      35  1.30 alive  ""    ""    ""    ""    ""   
#> 2 11    127894      1 swars2      44  1.30 alive  ""    ""    ""    ""    ""   
#> 3 11    127895      1 gustsu     215  1.30 alive  ""    ""    ""    ""    ""   
#> 4 11    127896      1 des2pa      32  1.30 alive  ""    ""    ""    ""    ""   
#> 5 11    127897      1 faraoc      70  1.30 alive  ""    ""    ""    ""    ""   
#> 6 11    127899      1 hybapr      31  1.30 alive  ""    ""    ""    ""    ""   
#> 
#> $`4917`
#> # A tibble: 6 x 12
#>   sq    tag    stemid mnemonic   dbh   hom status .dbh  code  .hom  note  comm 
#>   <chr> <chr>   <int> <chr>    <int> <dbl> <chr>  <chr> <chr> <chr> <chr> <chr>
#> 1 11    000046      1 gustsu     318  1.30 alive  ""    ""    ""    ""    ""   
#> 2 11    000047      1 alsebl     443  3.20 alive  ""    ""    ""    ""    ""   
#> 3 11    128097     10 hybapr      23  1.30 alive  ""    ""    ""    ""    ""   
#> 4 11    128097     11 hybapr      18  1.30 alive  ""    ""    ""    ""    ""   
#> 5 11    128098      1 gustsu     205  1.30 alive  ""    ""    ""    ""    ""   
#> 6 11    128100      4 hybapr      23  1.30 alive  ""    ""    ""    ""    ""   
#> 
#> $`4918`
#> # A tibble: 6 x 12
#>   sq    tag    stemid mnemonic   dbh   hom status .dbh  code  .hom  note  comm 
#>   <chr> <chr>   <int> <chr>    <int> <dbl> <chr>  <chr> <chr> <chr> <chr> <chr>
#> 1 11    000038      1 termob     574  3.20 alive  ""    ""    ""    ""    ""   
#> 2 11    128292      1 hirttr     233  1.30 alive  ""    ""    ""    ""    ""   
#> 3 11    128294      1 cappfr      35  1.30 alive  ""    ""    ""    ""    ""   
#> 4 11    128295      1 protte      24  1.30 alive  ""    ""    ""    ""    ""   
#> 5 11    128297      1 hybapr      28  1.30 alive  ""    ""    ""    ""    ""   
#> 6 11    128298      1 tachve      NA NA    dead   ""    ""    ""    ""    ""   
#> 


# You may customize the prepared data before making the fielforms. For example:
# Making names uppercase
prep <- lapply(prep, function(x) setNames(x, toupper(names(x))))
# Renaming a column
prep <- lapply(prep, function(x) dplyr::rename(x, COMMENTS = COMM))
# Removing a column
prep <- lapply(prep, function(x) dplyr::select(x, -NOTE))
# Adding a column
prep <- lapply(prep, function(x) dplyr::mutate(x, NEW = " "))

# Now you can make the fieldforms
tmp <- tempdir()
fieldforms_output(prep, output_dir = tmp)

#> 
#> 
#> processing file: forms_loop.Rmd

#>   |                                                                         |                                                                 |   0%  |                                                                         |......................                                           |  33%
#>   ordinary text without R code
#> 
#>   |                                                                         |...........................................                      |  67%
#> label: unnamed-chunk-1 (with options) 
#> List of 2
#>  $ results: chr "asis"
#>  $ echo   : logi FALSE
#> 
#>   |                                                                         |.................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: forms_loop.knit.md

#> "C:/Users/LeporeM/AppData/Local/Pandoc/pandoc" +RTS -K512m -RTS forms_loop.utf8.md --to docx --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash+smart --output pandoc3dc05fc27656.docx --highlight-style tango --reference-doc reference.docx 

#> 
#> Output created: C:/Users/LeporeM/AppData/Local/Temp/2/RtmpWuCzTS/fieldforms.docx

fieldforms_output(
  prep, "fieldforms2.docx", header = "Custom Header", output_dir = tmp
)

#> 
#> 
#> processing file: forms_loop.Rmd

#>   |                                                                         |                                                                 |   0%  |                                                                         |......................                                           |  33%
#>   ordinary text without R code
#> 
#>   |                                                                         |...........................................                      |  67%
#> label: unnamed-chunk-1 (with options) 
#> List of 2
#>  $ results: chr "asis"
#>  $ echo   : logi FALSE
#> 
#>   |                                                                         |.................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: forms_loop.knit.md

#> "C:/Users/LeporeM/AppData/Local/Pandoc/pandoc" +RTS -K512m -RTS forms_loop.utf8.md --to docx --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash+smart --output pandoc3dc048a54293.docx --highlight-style tango --reference-doc reference.docx 

#> 
#> Output created: C:/Users/LeporeM/AppData/Local/Temp/2/RtmpWuCzTS/fieldforms2.docx


# Confirm
dir(tmp)[grepl("docx$", dir(tmp))]

#> [1] "fieldforms.docx"  "fieldforms2.docx"


# Cleaning temporary directory
unlink(tmp)