Read all .csv or flat excel files from a directory into dataframes.

Read all .csv (csv_*()), or excel files (xl_*()) from a directory into a single dataframe (*_df()) or a list of dataframes (*_dfs()) -- where each element of the list is named as the source file. Excel files are treated as if they were flat -- meaning that these xl_*() functions can read only a single sheet (specified via the argument sheet passed to readxl::read_excel() via ...). For reading multiple all sheets in a single file see xlsheets_to_dfs().

csv_to_df(dir, ...)

csv_to_dfs(dir, ...)

xl_to_df(dir, ...)

xl_to_dfs(dir, ...)

Arguments

dir
String giving the directory containing the excel workbooks to read from.

...
Arguments passed to readr::read_csv() (csv_*()) or readxl::read_excel() (xl_*()).

Value

The versions ending in _df output a single dataframe. The versions ending in _dfs output a list of dataframes.

Acknowledgment

Thanks to Jessica Shue for inspiring this functions.

See also

Examples

xl_to_dfs(tool_example("multiple_workbooks"))

#> $workbook1.xlsx
#> # A tibble: 3 x 2
#>   numbers letters
#>     <dbl> <chr>  
#> 1       1 a      
#> 2       2 b      
#> 3       3 c      
#> 
#> $workbook2.xlsx
#> # A tibble: 3 x 2
#>   numbers letters
#>     <dbl> <chr>  
#> 1       1 a      
#> 2       2 b      
#> 3       3 c      
#> 

xl_to_df(tool_example("multiple_workbooks"))

#> # A tibble: 6 x 2
#>   numbers letters
#>     <dbl> <chr>  
#> 1       1 a      
#> 2       2 b      
#> 3       3 c      
#> 4       1 a      
#> 5       2 b      
#> 6       3 c      

# Pass an argument to `read_excel()` via `...`
xl_to_dfs(tool_example("multiple_workbooks"), sheet = 2)

#> $workbook1.xlsx
#> # A tibble: 3 x 2
#>   colors LETTERS
#>   <chr>  <chr>  
#> 1 red    A      
#> 2 yellow B      
#> 3 blue   C      
#> 
#> $workbook2.xlsx
#> # A tibble: 3 x 2
#>   colors LETTERS
#>   <chr>  <chr>  
#> 1 red    A      
#> 2 yellow B      
#> 3 blue   C      
#> 

csv_to_dfs(tool_example("multiple_csv"))

#> Parsed with column specification:
#> cols(
#>   numbers = col_integer(),
#>   letters = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   numbers = col_integer(),
#>   letters = col_character()
#> )

#> $csv1.csv
#> # A tibble: 3 x 2
#>   numbers letters
#>     <int> <chr>  
#> 1       1 a      
#> 2       2 b      
#> 3       3 c      
#> 
#> $csv2.csv
#> # A tibble: 3 x 2
#>   numbers letters
#>     <int> <chr>  
#> 1       4 a      
#> 2       5 b      
#> 3       6 c      
#> 

csv_to_df(tool_example("multiple_csv"))

#> Parsed with column specification:
#> cols(
#>   numbers = col_integer(),
#>   letters = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   numbers = col_integer(),
#>   letters = col_character()
#> )

#> # A tibble: 6 x 2
#>   numbers letters
#>     <int> <chr>  
#> 1       1 a      
#> 2       2 b      
#> 3       3 c      
#> 4       4 a      
#> 5       5 b      
#> 6       6 c      

# Pass an argument to `read_csv()` via `...`
csv_to_dfs(tool_example("multiple_csv"), n_max = 2)

#> Parsed with column specification:
#> cols(
#>   numbers = col_integer(),
#>   letters = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   numbers = col_integer(),
#>   letters = col_character()
#> )

#> $csv1.csv
#> # A tibble: 2 x 2
#>   numbers letters
#>     <int> <chr>  
#> 1       1 a      
#> 2       2 b      
#> 
#> $csv2.csv
#> # A tibble: 2 x 2
#>   numbers letters
#>     <int> <chr>  
#> 1       4 a      
#> 2       5 b      
#>