Functions to get variables from other variables.

These functios wrap their corresponding functions from the CTFS R Package, but these versions are stricter. The main differences are these:

names use "_" not ".".

argument gridsize defaults to missing to force the user to provide it.

If the argument plotdim is missing from functions gxgy_*(), its value will be guessed and notified.

rowcol_to_index(rowno, colno, gridsize, plotdim)

index_to_rowcol(index, gridsize, plotdim)

gxgy_to_index(gx, gy, gridsize, plotdim)

gxgy_to_lxly(gx, gy, gridsize, plotdim)

gxgy_to_rowcol(gx, gy, gridsize, plotdim)

gxgy_to_hectindex(gx, gy, plotdim)

Arguments

rowno, colno
Row and column number -- as defined in a census plot.

gridsize
The gridsize of the census plot (commonly 20 m).

plotdim
The global dimensions of the census plot (i.e. the maximum possible values of `gx` and `gy`).

index
Index number as defined for a census plot.

gx, gy
A number; golbal x and y position in a census plot.

Value

A vector or dataframe (see examples).

Details

Best is to avoid these functions because some of them output a vector and others output a dataframe. The recomended alternative to all these functions is the wrapper add_var(), which consistently inputs and outputs a dataframe, and is formally tested.

See also

add_var().

Examples

# NOT RUN {
gxgy_to_index(c(0, 400, 990), c(0, 200, 490), gridsize = 20)

gridsize <- 20
plotdim <- c(1000, 500)

x <- gxgy_to_hectindex(1:3, 1:3, plotdim)
x
typeof(x)
is.data.frame(x)
is.vector(x)

x <- gxgy_to_index(1:3, 1:3, gridsize, plotdim)
x
typeof(x)
is.data.frame(x)
is.vector(x)

x <- gxgy_to_lxly(1:3, 1:3, gridsize, plotdim)
x
typeof(x)
is.data.frame(x)
is.vector(x)

x <- gxgy_to_rowcol(1:3, 1:3, gridsize, plotdim)
x
typeof(x)
is.data.frame(x)
is.vector(x)

x <- index_to_rowcol(1:3, gridsize, plotdim)
x
typeof(x)
is.data.frame(x)
is.vector(x)

x <- rowcol_to_index(1:3, 1:3, gridsize, plotdim)
x
typeof(x)
is.data.frame(x)
is.vector(x)
# }