Create tidy names that are lowercase and have no empty spaces.

These functions create tidy names that are lowercase and have no empty spaces. nms_tidy() tidyes the names of named objects. Unnamed strings are also tidied via as_tidy_names(). as_tidy_names() tidies a string -- not its names.

nms_tidy(x)

as_tidy_names(x)

Arguments

x
A named object or a character string.

Value

A modified version of x with tidy names or a string of tidy names.

Examples

messy <- "Hi mE"

nms_tidy(messy)

#> [1] "hi_me"

# Same
as_tidy_names(messy)

#> [1] "hi_me"

messy_df <- setNames(data.frame(1), messy)
messy_df

#>   Hi mE
#> 1     1

nms_tidy(messy_df)

#>   hi_me
#> 1     1

messy_list <- setNames(list(1), messy)
messy_list

#> $`Hi mE`
#> [1] 1
#> 

nms_tidy(messy_list)

#> $hi_me
#> [1] 1
#> 

messy_vector <- setNames(c(1), messy)
messy_vector

#> Hi mE 
#>     1 

nms_tidy(messy_vector)

#> hi_me 
#>     1