Create tidy names that are lowercase and have no empty spaces.

These functions create tidy names that are lowercase and have no empty spaces. nms_tidy() tidies the names of named objects. Unnamed strings are also tidied via to_tidy_names(). to_tidy_names() tidies a string -- not its names.

nms_tidy(x)

Arguments

x
A named object or a character string.

Value

A modified version of x with tidy names or a string of tidy names.

Examples

# Use to_tidy_names()
library(fgeo.base)

messy <- "Hi yOu"

# With unnamed strings, both functions do the same
unnamed_string <- messy
names(messy)

#> NULL

nms_tidy(messy)

#> [1] "hi_you"

# Same
to_tidy_names(messy)

#> [1] "hi_you"

# WHY TWO FUNCTIONS?

messy_named_string <- c(`Messy Name` = messy)
# Targets names
nms_tidy(messy_named_string)

#> messy_name 
#>   "Hi yOu" 

# Targets strings -- not its names
to_tidy_names(messy_named_string)

#> Messy Name 
#>   "hi_you" 

# Same output, but here `to_tidy_names()` better communicates intention.
dfm <- data.frame(1)
setNames(dfm, nms_tidy(messy))

#>   hi_you
#> 1      1

setNames(dfm, to_tidy_names(messy))

#>   hi_you
#> 1      1

# Makes more sense when operating on strings
setNames(list(1), to_tidy_names(messy))

#> $hi_you
#> [1] 1
#> 

# Makes more sense when operating on named objects
messy_list <- list(`Hi yOu` = 1)
nms_tidy(messy_list)

#> $hi_you
#> [1] 1
#>