Pick and drop rows of a ForestGEO ViewFullTable or census table.

These functions provide an expressive and convenient way to filter ForestGEO's data. They let you remove missing values (with na.rm = TRUE) but conservatively default to keeping NAs. This behavior is similar to base subsetting and unlike dplyr::filter(); this conservative default matters because the difference is important between missing trees and dead trees -- and you may want to include missing trees in your analysis.

pick_dbh_min(dfm, x, na.rm = FALSE)

pick_dbh_max(dfm, x, na.rm = FALSE)

pick_dbh_under(dfm, x, na.rm = FALSE)

pick_dbh_over(dfm, x, na.rm = FALSE)

pick_status(dfm, x, na.rm = FALSE)

drop_status(dfm, x, na.rm = FALSE)

Arguments

dfm
Dataframe; A ForestGEO stem- or tree-table.

x
An atomic vector; a single value against to compare each value of the variable encoded in the function name.

na.rm
Set to TRUE if you want to remove missing values from the variable encoded in the function name.

Value

Dataframe rows with matching conditions.

See also

Examples

cns <- data.frame(
  dbh = c(0, 50, 100, 150, NA, NA, NA),
  status = c(rep("A", 4), "M", "D", NA)
  # stringsAsFactors = FALSE
)
cns

#>   dbh status
#> 1   0      A
#> 2  50      A
#> 3 100      A
#> 4 150      A
#> 5  NA      M
#> 6  NA      D
#> 7  NA   <NA>

# <=
pick_dbh_max(cns, 100)

#>   dbh status
#> 1   0      A
#> 2  50      A
#> 3 100      A
#> 5  NA      M
#> 6  NA      D
#> 7  NA   <NA>

pick_dbh_max(cns, 100, na.rm = TRUE)

#>   dbh status
#> 1   0      A
#> 2  50      A
#> 3 100      A

# >=
pick_dbh_min(cns, 100)

#>   dbh status
#> 3 100      A
#> 4 150      A
#> 5  NA      M
#> 6  NA      D
#> 7  NA   <NA>

pick_dbh_min(cns, 100, na.rm = TRUE)

#>   dbh status
#> 3 100      A
#> 4 150      A

# <
pick_dbh_under(cns, 100)

#>   dbh status
#> 1   0      A
#> 2  50      A
#> 5  NA      M
#> 6  NA      D
#> 7  NA   <NA>

pick_dbh_under(cns, 100, na.rm = TRUE)

#>   dbh status
#> 1   0      A
#> 2  50      A

# >
pick_dbh_over(cns, 100)

#>   dbh status
#> 4 150      A
#> 5  NA      M
#> 6  NA      D
#> 7  NA   <NA>

pick_dbh_over(cns, 100, na.rm = TRUE)

#>   dbh status
#> 4 150      A

# Same, but `subset()` does not let you keep NAs.
subset(cns, dbh > 100)

#>   dbh status
#> 4 150      A

# ==
pick_status(cns, "A")

#>   dbh status
#> 1   0      A
#> 2  50      A
#> 3 100      A
#> 4 150      A
#> 7  NA   <NA>

pick_status(cns, "A", na.rm = TRUE)

#>   dbh status
#> 1   0      A
#> 2  50      A
#> 3 100      A
#> 4 150      A

# !=
drop_status(cns, "D")

#>   dbh status
#> 1   0      A
#> 2  50      A
#> 3 100      A
#> 4 150      A
#> 5  NA      M
#> 7  NA   <NA>

drop_status(cns, "D", na.rm = TRUE)

#>   dbh status
#> 1   0      A
#> 2  50      A
#> 3 100      A
#> 4 150      A
#> 5  NA      M

# Compose
pick_dbh_over(drop_status(cns, "D", na.rm = TRUE), 100)

#>   dbh status
#> 4 150      A
#> 5  NA      M