Pick the main stem or main stemid(s) of each tree in each census.

pick_main_stem()picks the main stem of each tree in each census. It collapses data of multi-stem trees by picking a single stem per treeid per censusid: within this group, it picks the stem at the top of a list sorted first by descending order of hom and then by descending order of dbh -- this corrects the effect of buttresses and picks the main stem. It ignores groups of grouped data and rejects data with multiple plots (to work with data with multiple plotnames you may use split() or dplyr::nest()).

pick_main_stemid() does one step less than `pick_main_stem(). It only picks the main stemid(s) of each tree in each census and keeps all stems per treeid. This is useful when calculating the total basal area of a tree, because you need to sum the basal area of each individual stem as well as sum only one of the potentially multiple measurements of each buttressed stem per census.

pick_main_stem(.x)

pick_main_stemid(.x)

Arguments

.x
A ForestGEO-like dataframe, census or ViewFullTable.

Value

A dataframe with one row per per treeid per censusid and a single plotname.

Warning

These functions may be considerably slow. They are fastest if the data already has a single stem per treeid. They are slower with data containing multiple stems per treeid (per censusid) -- this is the main reason for using this function. The slowest scenario is when data also contains duplicated values of stemid per treeid (per censusid) -- this may happen if trees have buttresses -- in which case, these functions will check every stem for potential duplicates and pick the one with the largest hom value. In my computer, for example, a dataset of 2 million rows with multiple stems and buttresses took about 3 minutes to run, whereas a dataset with 2 million rows made up entirely of main stems took about ten seconds to run.

See also

Examples

library(tibble)

# Trees with buttresses may have more than one measurements per stem.
census <- tribble(
    ~sp, ~treeID, ~stemID,  ~hom, ~dbh, ~CensusID,
  "sp1",     "1",   "1.1",   140,   40,         1,  # main stem
  "sp1",     "1",   "1.1",   130,   60,         1,
  "sp1",     "1",   "1.2",   130,   55,         1,
  "sp2",     "2",   "2.1",   130,    5,         1   # main stem
)

# Picks largest hom first (to correct effect of batreesses) then largest dbh
pick_main_stem(census)

#> # A tibble: 2 x 6
#>   sp    treeID stemID   hom   dbh CensusID
#>   <chr> <chr>  <chr>  <dbl> <dbl>    <dbl>
#> 1 sp1   1      1.1      140    40        1
#> 2 sp2   2      2.1      130     5        1

# Picks the main stemid of each stem and keeps all stems of each tree.
pick_main_stemid(census)

#> # A tibble: 3 x 6
#>   sp    treeID stemID   hom   dbh CensusID
#>   <chr> <chr>  <chr>  <dbl> <dbl>    <dbl>
#> 1 sp1   1      1.1      140    40        1
#> 2 sp1   1      1.2      130    55        1
#> 3 sp2   2      2.1      130     5        1