Pick main stems within a specific dbh range.

These functions extend pick_main_stem(). Working with simple or grouped dataframes (via dplyr::group_by()) these functions let you pick main stems within a given dbh range:

pick_woods() is a general function and lets you choose the dbh range. The other functions are shortcuts with pre-defined dbh ranges:

pick_saplings_and_trees() picks stems of 10 mm dbh and above.

pick_saplings() picks stems between 10 mm dbh inclusive and 100 mm dbh exclusive.

pick_trees() picks stems of 100 mm dbh and above.

pick_woods(.data, ...)

pick_trees(.data)

pick_saplings(.data)

pick_saplings_and_trees(.data)

Arguments

.data
A ForestGEO-like dataframe, either a census or ViewFullTable.

...
Expressions passed to dplyr::filter(), E.g. the expression dbh >= 100 picks dbh values of 100 mm and above.

Value

A dataframe. See description.

Examples

library(tibble)

census <- tribble(
    ~sp, ~treeID, ~stemID,  ~hom, ~dbh, ~CensusID,
  "sp1",     "1",   "1.1",   140,   40,         1,  # main stem
  "sp1",     "1",   "1.1",   130,   60,         1   # buttress
)

# Trees with buttresses may have more than one measurements per stem.
# Piks largest hom first (to correct effect of batreesses) then largest dbh
pick_main_stem(census)

#> # A tibble: 1 x 6
#>   sp    treeID stemID   hom   dbh CensusID
#>   <chr> <chr>  <chr>  <dbl> <dbl>    <dbl>
#> 1 sp1   1      1.1      140    40        1

# Also works with multiple censuses
censuses <- tribble(
    ~sp, ~treeID, ~stemID,  ~hom, ~dbh, ~CensusID,
  "sp1",     "1",   "1.1",   140,   40,         1,  # main stem
  "sp1",     "1",   "1.1",   130,   60,         1,  # buttress
  "sp1",     "1",   "1.1",   140,   50,         2,  # main stem
  "sp1",     "1",   "1.1",   130,   70,         2   # buttress
)

pick_main_stem(censuses)

#> # A tibble: 2 x 6
#>   sp    treeID stemID   hom   dbh CensusID
#>   <chr> <chr>  <chr>  <dbl> <dbl>    <dbl>
#> 1 sp1   1      1.1      140    40        1
#> 2 sp1   1      1.1      140    50        2

# pick_woods() and friends internally pick the main stem, then pick by dbh
pick_woods(censuses)

#> # A tibble: 2 x 6
#>   sp    treeID stemID   hom   dbh CensusID
#>   <chr> <chr>  <chr>  <dbl> <dbl>    <dbl>
#> 1 sp1   1      1.1      140    40        1
#> 2 sp1   1      1.1      140    50        2

pick_woods(censuses, dbh > 40)

#> # A tibble: 1 x 6
#>   sp    treeID stemID   hom   dbh CensusID
#>   <chr> <chr>  <chr>  <dbl> <dbl>    <dbl>
#> 1 sp1   1      1.1      140    50        2

# These functions are convenient shortcuts
census <- tribble(
    ~sp, ~treeID, ~stemID,  ~hom, ~dbh, ~CensusID,
  "sp1",     "1",   "1.1",   140,   40,         1,  # main stem
  "sp1",     "1",   "1.1",   130,   60,         1,  # buttress
  "sp1",     "2",   "2.1",   130,   100,        1,
  "sp1",     "3",   "2.1",   130,   110,        1,
)

pick_saplings(census)

#> # A tibble: 1 x 6
#>   sp    treeID stemID   hom   dbh CensusID
#>   <chr> <chr>  <chr>  <dbl> <dbl>    <dbl>
#> 1 sp1   1      1.1      140    40        1

pick_trees(census)

#> # A tibble: 2 x 6
#>   sp    treeID stemID   hom   dbh CensusID
#>   <chr> <chr>  <chr>  <dbl> <dbl>    <dbl>
#> 1 sp1   2      2.1      130   100        1
#> 2 sp1   3      2.1      130   110        1

pick_saplings_and_trees(census)

#> # A tibble: 3 x 6
#>   sp    treeID stemID   hom   dbh CensusID
#>   <chr> <chr>  <chr>  <dbl> <dbl>    <dbl>
#> 1 sp1   1      1.1      140    40        1
#> 2 sp1   2      2.1      130   100        1
#> 3 sp1   3      2.1      130   110        1