Import multiple .rdata files from a directory into a dataframe.

This is shortcut for list_df(rdata_list(x)) with the additional ability to read exclusively .rdata files containing a specific string. It is designed to expect that all .rdata files contain dataframes with the same names -- as is the case for multiple ForestGEO censuses delivered by the database. Notice that function name have the format input_output, i.e. rdata_df.

rdata_df(path_dir, match = NULL, .id = NULL)

Arguments

path_dir
String; the path to a directory containing the files to read (all must be of appropriate format; see examples).

match
If not NULL a sting to match specific .rdata file-names (excluding the extension).

.id
If not NULL a variable with this name will be created to inform the source file.

Value

A dataframe.

See also

Examples

path_dir <- tool_example("rdata")
dir(path_dir)

#> [1] "tree5.RData" "tree6.RData"

rdata_df(path_dir)

#> # A tibble: 2,000 x 19
#>    treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID
#>     <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int>
#>  1    104    143 10009 10009   DACE~ 113      10.3  245.    439947        5
#>  2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    466597        5
#>  3    180    225 1001~ 100174  CASA~ 921     165.   410.    466623        5
#>  4    602    736 1006~ 100649  GUAG~ 821     149.   414.    466727        5
#>  5    631    775 10069 10069   PREM~ 213      38.3  245.    439989        5
#>  6    647    793 1007~ 100708  SCHM~ 821     143.   411.    466743        5
#>  7   1086   1339 10122 10122   DRYG~ 413      68.9  253.    440021        5
#>  8   1144   1410 1012~ 101285  SCHM~ 920     161.   395.    466889        5
#>  9   1168   1438 10131 10131   DACE~ 413      70.6  252.    440031        5
#> 10   1380 114352 1015~ 149529  CASA~ 820     142.   386.    466957        5
#> # ... with 1,990 more rows, and 9 more variables: dbh <dbl>, pom <chr>,
#> #   hom <dbl>, ExactDate <date>, DFstatus <chr>, codes <chr>, nostems <dbl>,
#> #   status <chr>, date <dbl>

rdata_df(path_dir, match = "tree6")

#> # A tibble: 1,000 x 19
#>    treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID
#>     <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int>
#>  1    104    143 10009 10009   DACE~ 113      10.3  245.    582850        6
#>  2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    578696        6
#>  3    180    225 1001~ 100174  CASA~ 921     165.   410.    617049        6
#>  4    602    736 1006~ 100649  GUAG~ 821     149.   414.    614253        6
#>  5    631    775 10069 10069   PREM~ 213      38.3  245.    598429        6
#>  6    647    793 1007~ 100708  SCHM~ 821     143.   411.    614211        6
#>  7   1086   1339 10122 10122   DRYG~ 413      68.9  253.    603131        6
#>  8   1144   1410 1012~ 101285  SCHM~ 920     161.   395.    616923        6
#>  9   1168   1438 10131 10131   DACE~ 413      70.6  252.    603151        6
#> 10   1380   1702 1015~ 101560  CASA~ 820     142.   386.    614023        6
#> # ... with 990 more rows, and 9 more variables: dbh <dbl>, pom <chr>,
#> #   hom <dbl>, ExactDate <date>, DFstatus <chr>, codes <chr>, nostems <dbl>,
#> #   status <chr>, date <dbl>

dfm <- rdata_df(path_dir, match = "tree5|6", .id = "source")
dfm

#> # A tibble: 2,000 x 20
#>    source treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID
#>    <chr>   <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>
#>  1 tree5     104    143 10009 10009   DACE~ 113      10.3  245.    439947
#>  2 tree5     119    158 1001~ 100104  MYRS~ 1021    183.   410.    466597
#>  3 tree5     180    225 1001~ 100174  CASA~ 921     165.   410.    466623
#>  4 tree5     602    736 1006~ 100649  GUAG~ 821     149.   414.    466727
#>  5 tree5     631    775 10069 10069   PREM~ 213      38.3  245.    439989
#>  6 tree5     647    793 1007~ 100708  SCHM~ 821     143.   411.    466743
#>  7 tree5    1086   1339 10122 10122   DRYG~ 413      68.9  253.    440021
#>  8 tree5    1144   1410 1012~ 101285  SCHM~ 920     161.   395.    466889
#>  9 tree5    1168   1438 10131 10131   DACE~ 413      70.6  252.    440031
#> 10 tree5    1380 114352 1015~ 149529  CASA~ 820     142.   386.    466957
#> # ... with 1,990 more rows, and 10 more variables: CensusID <int>, dbh <dbl>,
#> #   pom <chr>, hom <dbl>, ExactDate <date>, DFstatus <chr>, codes <chr>,
#> #   nostems <dbl>, status <chr>, date <dbl>

tail(dfm)

#> # A tibble: 6 x 20
#>   source treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID
#>   <chr>   <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>
#> 1 tree6  126869 166460 1786~ 178613  CASA~ 1122     211.  420.    619723
#> 2 tree6  126934 166531 1787~ 178763  MICR~ 324       47   480.    619794
#> 3 tree6  127015 166617 1793~ 179345  CLUR~ 1025     186.  485.    619880
#> 4 tree6  127017 166619 1793~ 179361  PSYB~ 1025     185.  494.    619882
#> 5 tree6  127025 166627 1794~ 179417  TETB~ 623      112.  447.    619890
#> 6 tree6  127033 166635 1794~ 179463  TETB~ 1023     184.  454.    619898
#> # ... with 10 more variables: CensusID <int>, dbh <dbl>, pom <chr>, hom <dbl>,
#> #   ExactDate <date>, DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>,
#> #   date <dbl>