Read ForestGEO censuses.

This function reads any number of ForestGEO censuses stored in .rdata files and stores them in a nested dataframe of class 'censuses_df', which has a convenient print method and enables using specific methods for working with ForestGEO data.

read_censuses(path_dir, .match = NULL, .id = "censuses")

Arguments

path_dir
String; the path to a directory containing the files to read (all must be of appropriate format; see examples).

.match
If not NULL a string to match specific .rdata file-names (excluding the extension).

.id
If not NULL a variable with this name will be created to inform the source file.

Value

An object of class 'censuses_df'.

See also

Examples

library(dplyr)

censuses <- read_censuses(tool_example("rdata"))
class(censuses)

#> [1] "censuses_df" "tbl_df"      "tbl"         "data.frame" 

censuses$data

#> [[1]]
#> # A tibble: 1,000 x 19
#>    treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID
#>     <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int>
#>  1    104    143 10009 10009   DACE~ 113      10.3  245.    439947        5
#>  2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    466597        5
#>  3    180    225 1001~ 100174  CASA~ 921     165.   410.    466623        5
#>  4    602    736 1006~ 100649  GUAG~ 821     149.   414.    466727        5
#>  5    631    775 10069 10069   PREM~ 213      38.3  245.    439989        5
#>  6    647    793 1007~ 100708  SCHM~ 821     143.   411.    466743        5
#>  7   1086   1339 10122 10122   DRYG~ 413      68.9  253.    440021        5
#>  8   1144   1410 1012~ 101285  SCHM~ 920     161.   395.    466889        5
#>  9   1168   1438 10131 10131   DACE~ 413      70.6  252.    440031        5
#> 10   1380 114352 1015~ 149529  CASA~ 820     142.   386.    466957        5
#> # ... with 990 more rows, and 9 more variables: dbh <dbl>, pom <chr>,
#> #   hom <dbl>, ExactDate <date>, DFstatus <chr>, codes <chr>, nostems <dbl>,
#> #   status <chr>, date <dbl>
#> 
#> [[2]]
#> # A tibble: 1,000 x 19
#>    treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID
#>     <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int>
#>  1    104    143 10009 10009   DACE~ 113      10.3  245.    582850        6
#>  2    119    158 1001~ 100104  MYRS~ 1021    183.   410.    578696        6
#>  3    180    225 1001~ 100174  CASA~ 921     165.   410.    617049        6
#>  4    602    736 1006~ 100649  GUAG~ 821     149.   414.    614253        6
#>  5    631    775 10069 10069   PREM~ 213      38.3  245.    598429        6
#>  6    647    793 1007~ 100708  SCHM~ 821     143.   411.    614211        6
#>  7   1086   1339 10122 10122   DRYG~ 413      68.9  253.    603131        6
#>  8   1144   1410 1012~ 101285  SCHM~ 920     161.   395.    616923        6
#>  9   1168   1438 10131 10131   DACE~ 413      70.6  252.    603151        6
#> 10   1380   1702 1015~ 101560  CASA~ 820     142.   386.    614023        6
#> # ... with 990 more rows, and 9 more variables: dbh <dbl>, pom <chr>,
#> #   hom <dbl>, ExactDate <date>, DFstatus <chr>, codes <chr>, nostems <dbl>,
#> #   status <chr>, date <dbl>
#> 

censuses %>%
  pick(dbh > 600) %>%
  pull(data)

#> [[1]]
#> # A tibble: 4 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1   1168   1438 10131 10131   DACE~ 413      70.6  252.    440031        5   604
#> 2  24399  30562 15172 15172   MANB~ 618     118.   343.    443041        5   777
#> 3  25821  32585 16675 16675   BUCT~ 1220    225.   397.    444015        5   955
#> 4  27376  34692 18275 18275   BUCT~ 1223    221.   445.    445060        5   661
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#> 
#> [[2]]
#> # A tibble: 4 x 19
#>   treeID stemID tag   StemTag sp    quadrat    gx    gy MeasureID CensusID   dbh
#>    <int>  <int> <chr> <chr>   <chr> <chr>   <dbl> <dbl>     <int>    <int> <dbl>
#> 1   1168   1438 10131 10131   DACE~ 413      70.6  252.    603151        6   613
#> 2  24399  30562 15172 15172   MANB~ 618     118.   343.    608417        6   812
#> 3  25821  32585 16675 16675   BUCT~ 1220    225.   397.    585671        6   992
#> 4  27376  34692 18275 18275   BUCT~ 1223    221.   445.    586086        6   676
#> # ... with 8 more variables: pom <chr>, hom <dbl>, ExactDate <date>,
#> #   DFstatus <chr>, codes <chr>, nostems <dbl>, status <chr>, date <dbl>
#>