Import ViewFullTable or ViewTaxonomy data from a .tsv or .csv file.

read_vft() and read_taxa() help you to read ViewFullTable and ViewTaxonomy data from text files delivered by the ForestGEO database. These functions avoid common problems about column separators, missing values, column names, and column types.

read_vft(file, delim = NULL, na = c("", "NA", "NULL"), ...)

read_taxa(file, delim = NULL, na = c("", "NA", "NULL"), ...)

Arguments

file
A path to a file.

delim
Single character used to separate fields within a record. The default (delim = NULL) is to guess between comma or tab ("," or "\t").

na
Character vector of strings to interpret as missing values. Set this option to character() to indicate no missing values.

...
Other arguments passed to readr::read_delim().

Value

A tibble.

Acknowledgments

Thanks to Shameema Jafferjee Esufali for inspiring the feature that automatically detects delim (issue 65).

See also

Examples

library(fgeo.x)

example_path()

#>  [1] "csv"           "mixed_files"   "rdata"         "rdata_one"    
#>  [5] "rds"           "taxa.csv"      "tsv"           "vft_4quad.csv"
#>  [9] "view"          "weird"         "xl"           

file_vft <- example_path("view/vft_4quad.csv")
read_vft(file_vft)

#> # A tibble: 500 x 32
#>     DBHID PlotName PlotID Family Genus SpeciesName Mnemonic Subspecies SpeciesID
#>     <int> <chr>     <int> <chr>  <chr> <chr>       <chr>    <chr>          <int>
#>  1 385164 luquillo      1 Rubia~ Psyc~ brachiata   PSYBRA   NA               185
#>  2 385261 luquillo      1 Urtic~ Cecr~ schreberia~ CECSCH   NA                74
#>  3 384600 luquillo      1 Rubia~ Psyc~ brachiata   PSYBRA   NA               185
#>  4 608789 luquillo      1 Rubia~ Psyc~ berteroana  PSYBER   NA               184
#>  5 388579 luquillo      1 Areca~ Pres~ acuminata   PREMON   NA               182
#>  6 384626 luquillo      1 Arali~ Sche~ morototoni  SCHMOR   NA               196
#>  7 410958 luquillo      1 Rubia~ Psyc~ brachiata   PSYBRA   NA               185
#>  8 385102 luquillo      1 Piper~ Piper glabrescens PIPGLA   NA               174
#>  9 353163 luquillo      1 Areca~ Pres~ acuminata   PREMON   NA               182
#> 10 481018 luquillo      1 Salic~ Case~ arborea     CASARB   NA                70
#> # ... with 490 more rows, and 23 more variables: SubspeciesID <chr>,
#> #   QuadratName <chr>, QuadratID <int>, PX <dbl>, PY <dbl>, QX <dbl>, QY <dbl>,
#> #   TreeID <int>, Tag <chr>, StemID <int>, StemNumber <int>, StemTag <int>,
#> #   PrimaryStem <chr>, CensusID <int>, PlotCensusNumber <int>, DBH <dbl>,
#> #   HOM <dbl>, ExactDate <date>, Date <int>, ListOfTSM <chr>, HighHOM <int>,
#> #   LargeStem <chr>, Status <chr>

file_taxa <- example_path("view/taxa.csv")
read_taxa(file_taxa)

#> # A tibble: 163 x 21
#>    ViewID SpeciesID SubspeciesID Family Mnemonic Genus SpeciesName Rank 
#>     <int>     <int> <chr>        <chr>  <chr>    <chr> <chr>       <chr>
#>  1      1        56 NA           Fabac~ AESAME   Aesc~ americana   NA   
#>  2      2        57 NA           Eupho~ ALCFLO   Alch~ floribunda  NA   
#>  3      3        58 NA           Eupho~ ALCLAT   Alch~ latifolia   NA   
#>  4      4        59 NA           Fabac~ ANDINE   Andi~ inermis     NA   
#>  5      5        60 NA           Rubia~ ANTOBT   Sten~ obtusifoli~ NA   
#>  6      6        61 NA           Myrsi~ ARDGLA   Ardi~ glauciflora NA   
#>  7      7        62 NA           Morac~ ARTALT   Arto~ altilis     NA   
#>  8      8        63 NA           Laura~ BEIPEN   Beil~ pendula     NA   
#>  9      9        64 NA           Solan~ BRUPOR   Brun~ portoricen~ NA   
#> 10     10        65 NA           Combr~ BUCTET   Buch~ tetraphylla NA   
#> # ... with 153 more rows, and 13 more variables: Subspecies <chr>,
#> #   Authority <chr>, IDLevel <chr>, subspMnemonic <chr>, subspAuthority <chr>,
#> #   FieldFamily <chr>, Lifeform <chr>, Description <chr>, wsg <dbl>,
#> #   wsglevel <chr>, ListOfOldNames <chr>, Specimens <chr>, Reference <chr>