Sanitize ViewFullTable and ViewTaxonomy.

These functions sanitize dataframes from ViewFullTable and ViewTaxonomy to achieve the same effect as reading with read_vft() and read_taxa(). The most common problems that these functions fix are these:

Ensure that each column has the correct type.

Ensure that missing values are represented with NAs -- not with the literal string "NULL".

sanitize_vft(x, col_types = type_vft(), na = c("", "NA", "NULL"), ...)

Arguments

vft
A dataframe; particularly a ForestGEO ViewFullTable (vft).

taxa
A dataframe; particularly a ForestGEO ViewTaxonomy (taxa).

Value

A dataframe.

Acknowledgments

Thanks to Shameema Jafferjee Esufali for motivating this functions.

See also

Examples

vft <- fgeo.data::luquillo_vft_4quad

# Introdue problems to show how to fix them
# Bad column types
vft[] <- lapply(vft, as.character)
# Bad representation of missing values
vft$PlotName <- "NULL"

# "NULL" should be replaced by `NA` and `DBH` should be numeric -- not character
str(vft[c("PlotName", "DBH")])

#> Classes 'tbl_df', 'tbl' and 'data.frame':	3302 obs. of  2 variables:
#>  $ PlotName: chr  "NULL" "NULL" "NULL" "NULL" ...
#>  $ DBH     : chr  NA "185" NA "341" ...

vft_sane <- sanitize_vft(vft)
str(vft_sane[c("PlotName", "DBH")])

#> Classes 'tbl_df', 'tbl' and 'data.frame':	3302 obs. of  2 variables:
#>  $ PlotName: chr  NA NA NA NA ...
#>  $ DBH     : num  NA 185 NA 341 370 375 167 172 187 141 ...

taxa <- fgeo.data::luquillo_taxa
# Bad column types
taxa[] <- lapply(taxa, as.character)
# Bad representation of missing values
taxa$SubspeciesID <- "NULL"

# "NULL" should be replaced by `NA` and `ViewID` should be integer
str(taxa[c("SubspeciesID", "ViewID")])

#> Classes 'tbl_df', 'tbl' and 'data.frame':	163 obs. of  2 variables:
#>  $ SubspeciesID: chr  "NULL" "NULL" "NULL" "NULL" ...
#>  $ ViewID      : chr  "1" "2" "3" "4" ...

taxa_sane <- sanitize_taxa(taxa)
str(taxa_sane[c("SubspeciesID", "ViewID")])

#> Classes 'tbl_df', 'tbl' and 'data.frame':	163 obs. of  2 variables:
#>  $ SubspeciesID: chr  NA NA NA NA ...
#>  $ ViewID      : int  1 2 3 4 5 6 7 8 9 10 ...