Combine spreadsheets from excel workbooks and output common data structures.

These functions combine spreadsheets from excel workbooks into common data structures. xl_sheets_to_csv() and xl_sheets_to_xl() write a .csv or excel (.xlsx) file per workbook -- combining all spreadsheets. xl_sheets_to_df outputs a list where each dataframes combines all spreadsheeets of a workbook.

xl_sheets_to_csv(input_dir, output_dir = "./", first_census = FALSE)

xl_sheets_to_xl(input_dir, output_dir = "./", first_census = FALSE)

xl_sheets_to_df(input_dir, first_census = FALSE)

Arguments

input_dir
String giving the directory containing the excel workbooks to read from.

output_dir
String giving the directory where to write .csv files to.

first_census
This argument tells these functions what sheets to expect in the input.

Use TRUE if this is your first census. The expected input must have sheets (1) "root", (2) "multi_stems", (3) "secondary_stems", and (4) "single_stems".

Use FALSE (default) if this is not your first census. The expected input must have sheets (1) "root", (2) "original_stems", (3) "new_secondary_stems", and (4) "recruits".

Value

Writes one .csv file for each workbook.

Details

This is a rigid function with a very specific goal: To process data from a specific sampling software -- FastField. Specifically, this is what this function does:

Reads each spreadsheet from each workbook and map it to a dataframe.

Lowercases and links the names of each dataframe.

Keeps only these dataframes: (1) original_stems, (2) new_secondary_stems, and (3) "recruits".

Dates the data by submission_id (date comes from the spreadsheet root).

Lowercases and links the names of each dataframe-variable.

Drops fake stems.

Output a common data structure of your choice.

Acknowledgment

Sabrina Russo helped to make these functions useful with first censuses.

Examples

library(fs)
library(readr)
library(readxl)

# NOT A FIRST CENSUS
# Path to the folder I want to read excel files from
input_dir <- dirname(example_path("two_files/new_stem_1.xlsx"))
input_dir

#> [1] "C:/Users/LeporeM/Dropbox/git_repos/fgeo.tool/inst/extdata/two_files"

# Files I want to read
dir(input_dir, pattern = "xlsx")

#> [1] "new_stem_1.xlsx" "new_stem_2.xlsx"

# Path to the folder I want to write .csv files to
output_dir <- tempdir()

# Output a csv file
xl_sheets_to_csv(input_dir, output_dir)

# Confirm
path_file(dir_ls(output_dir, regexp = "new_stem.*csv$"))

#> new_stem_0.csv new_stem_1.csv new_stem_2.csv 

# Also possible to output excel and a list of dataframe. See next section.

# FIRST CENSUS
input_dir <- dirname(example_path("first_census/census.xlsx"))
# As a reminder you'll get a warning of missing sheets
# Output list of dataframes (one per input workbook -- here only one)
xl_sheets_to_df(input_dir, first_census = TRUE)

#> Warning: `.x` has no dataframe new_secondary_stems

#> Warning: `.x` has no dataframe recruits

#> $census.xlsx
#>                           submission_id quadrat tag stem_tag
#> 1  c01b37d5-2c30-4086-b4a1-78327dfa0a9a    0816 007        1
#> 2  efd8e991-2e34-40c8-b503-44def6e542d9    0916 004        1
#> 3  efd8e991-2e34-40c8-b503-44def6e542d9    0916 007        1
#> 4  efd8e991-2e34-40c8-b503-44def6e542d9    0916 008        1
#> 5  efd8e991-2e34-40c8-b503-44def6e542d9    0916 010        1
#> 6  efd8e991-2e34-40c8-b503-44def6e542d9    0916 012        1
#> 7  7997596a-c94e-47b6-ab28-4b0e6a94e9df    1017 001        1
#> 8  572ddae1-d02c-4048-9eca-7c12454af24b    0816 002        1
#> 9  c01b37d5-2c30-4086-b4a1-78327dfa0a9a    0816 007        2
#> 10 c01b37d5-2c30-4086-b4a1-78327dfa0a9a    0816 007        3
#> 11 c01b37d5-2c30-4086-b4a1-78327dfa0a9a    0816 007        4
#> 12 c01b37d5-2c30-4086-b4a1-78327dfa0a9a    0816 007        5
#> 13 efd8e991-2e34-40c8-b503-44def6e542d9    0916 004        2
#> 14 efd8e991-2e34-40c8-b503-44def6e542d9    0916 004        3
#> 15 efd8e991-2e34-40c8-b503-44def6e542d9    0916 004        4
#> 16 efd8e991-2e34-40c8-b503-44def6e542d9    0916 004        5
#> 17 efd8e991-2e34-40c8-b503-44def6e542d9    0916 004        6
#> 18 efd8e991-2e34-40c8-b503-44def6e542d9    0916 004        7
#> 19 efd8e991-2e34-40c8-b503-44def6e542d9    0916 004        8
#> 20 efd8e991-2e34-40c8-b503-44def6e542d9    0916 007        2
#> 21 efd8e991-2e34-40c8-b503-44def6e542d9    0916 007        3
#> 22 efd8e991-2e34-40c8-b503-44def6e542d9    0916 007        4
#> 23 efd8e991-2e34-40c8-b503-44def6e542d9    0916 007        5
#> 24 efd8e991-2e34-40c8-b503-44def6e542d9    0916 008        2
#> 25 efd8e991-2e34-40c8-b503-44def6e542d9    0916 010        2
#> 26 efd8e991-2e34-40c8-b503-44def6e542d9    0916 012        2
#> 27 7997596a-c94e-47b6-ab28-4b0e6a94e9df    1017 001        2
#> 28 7997596a-c94e-47b6-ab28-4b0e6a94e9df    1017 001        3
#> 29 572ddae1-d02c-4048-9eca-7c12454af24b    0816 002        2
#> 30 c01b37d5-2c30-4086-b4a1-78327dfa0a9a    0816 006        1
#> 31 c01b37d5-2c30-4086-b4a1-78327dfa0a9a    0816 008        1
#> 32 c01b37d5-2c30-4086-b4a1-78327dfa0a9a    0816 009        1
#> 33 c01b37d5-2c30-4086-b4a1-78327dfa0a9a    0816 010        1
#> 34 efd8e991-2e34-40c8-b503-44def6e542d9    0916 001        1
#> 35 efd8e991-2e34-40c8-b503-44def6e542d9    0916 002        1
#> 36 efd8e991-2e34-40c8-b503-44def6e542d9    0916 003        1
#> 37 efd8e991-2e34-40c8-b503-44def6e542d9    0916 005        1
#> 38 efd8e991-2e34-40c8-b503-44def6e542d9    0916 006        1
#> 39 efd8e991-2e34-40c8-b503-44def6e542d9    0916 009        1
#> 40 efd8e991-2e34-40c8-b503-44def6e542d9    0916 011        1
#> 41 7997596a-c94e-47b6-ab28-4b0e6a94e9df    1017 002        1
#> 42 7997596a-c94e-47b6-ab28-4b0e6a94e9df    1017 003        1
#> 43 7997596a-c94e-47b6-ab28-4b0e6a94e9df    1017 004        1
#> 44 a4e952e6-dcad-4410-aba9-36918fa1fa62    <NA> 001        1
#> 45 a4e952e6-dcad-4410-aba9-36918fa1fa62    0818 002        1
#> 46 572ddae1-d02c-4048-9eca-7c12454af24b    0816 001        1
#> 47 572ddae1-d02c-4048-9eca-7c12454af24b    0816 003        1
#> 48 572ddae1-d02c-4048-9eca-7c12454af24b    0816 004        1
#> 49 572ddae1-d02c-4048-9eca-7c12454af24b    0816 005        1
#>                     species species_code  dbh status     codes           notes
#> 1  American Basswood/Linden       TILIAM 51.2     LI         M            <NA>
#> 2  American Basswood/Linden       TILIAM 31.6     LI         M            <NA>
#> 3                  Ironwood       OSTRVI 10.2     LI         M            <NA>
#> 4                  Ironwood       OSTRVI  2.7     LI       M;L   In brush pile
#> 5                 Hackberry       CELTOC 25.9     LI         M            <NA>
#> 6                   Bur Oak       QUERMA 35.2     LI   M;FS;FL            <NA>
#> 7  American Basswood/Linden       TILIAM 65.3     LI    M;FS;I            <NA>
#> 8                 Hackberry       CELTOC 45.1     LI         M            <NA>
#> 9                    TILIAM         <NA> 42.4     LI       S;L            <NA>
#> 10                   TILIAM         <NA> 39.8     LI         S            <NA>
#> 11                   TILIAM         <NA> 37.3     LI       S;L            <NA>
#> 12                   TILIAM         <NA> 38.5     LI       S;L            <NA>
#> 13                   TILIAM         <NA> 26.9     LI         S            <NA>
#> 14                   TILIAM         <NA>  1.3     LI         S            <NA>
#> 15                   TILIAM         <NA> 17.9     LI         S            <NA>
#> 16                   TILIAM         <NA>    1     LI         S            <NA>
#> 17                   TILIAM         <NA>    1     LI         S            <NA>
#> 18                   TILIAM         <NA>  1.3     LI         S            <NA>
#> 19                   TILIAM         <NA>  1.4     LI         S            <NA>
#> 20                   OSTRVI         <NA>  2.1     LI         S            <NA>
#> 21                   OSTRVI         <NA>  1.3     LI         S            <NA>
#> 22                   OSTRVI         <NA>  1.5     LI         S            <NA>
#> 23                   OSTRVI         <NA>  1.3     LI         S            <NA>
#> 24                   OSTRVI         <NA>  1.4     LI       S;L   In brush pile
#> 25                   CELTOC         <NA> 14.7     LI         S            <NA>
#> 26                   QUERMA         <NA> 31.5     LI S;FS;I;FL            <NA>
#> 27                   TILIAM         <NA> 43.2     LI    S;FS;I            <NA>
#> 28                   TILIAM         <NA> 24.5     LI    S;FS;I            <NA>
#> 29                   CELTOC         <NA> 27.5     LI         S            <NA>
#> 30                  Bur Oak       QUERMA 40.5     LI      <NA>            <NA>
#> 31                  Bur Oak       QUERMA   33     LI      <NA>            <NA>
#> 32                 Ironwood       OSTRVI  2.2     LI         L            <NA>
#> 33                 Ironwood       OSTRVI  1.8     LI         L       Map later
#> 34                  Bur Oak       QUERMA 29.4     LI         T            <NA>
#> 35                  Bur Oak       QUERMA 51.2     LI      <NA>            <NA>
#> 36                 Ironwood       OSTRVI  3.3     LI      <NA>            <NA>
#> 37                Hackberry       CELTOC 20.6     LI      <NA>            <NA>
#> 38                  Bur Oak       QUERMA 36.3     LI      <NA> Large root ball
#> 39                Hackberry       CELTOC 11.4     LI      <NA>            <NA>
#> 40                Hackberry       CELTOC 22.5     LI      <NA>            <NA>
#> 41                Hackberry       CELTOC 49.4     LI      FS;I            <NA>
#> 42                  Bur Oak       QUERMA 47.6     LI        FS            <NA>
#> 43                  Bur Oak       QUERMA 41.7     LI      <NA>  Dead Secondary
#> 44                  Bur Oak       QUERMA 52.5     LI     FS;FL            <NA>
#> 45                  Bur Oak       QUERMA 36.2     LI     FL;FS            <NA>
#> 46                  Bur Oak       QUERMA 39.8     LI      <NA>            <NA>
#> 47                  Bur Oak       QUERMA 25.8     LI      <NA>            <NA>
#> 48                  Bur Oak       QUERMA 40.5     LI      <NA>            <NA>
#> 49                Hackberry       CELTOC   42     LI      <NA>            <NA>
#>    pom           sheet section_id unique_stem       date
#> 1  1.3     multi_stems       <NA>       007_1 2018.05.14
#> 2  1.3     multi_stems       <NA>       004_1 2018.05.14
#> 3  1.3     multi_stems       <NA>       007_1 2018.05.14
#> 4  1.3     multi_stems       <NA>       008_1 2018.05.14
#> 5  1.3     multi_stems       <NA>       010_1 2018.05.14
#> 6  1.3     multi_stems       <NA>       012_1 2018.05.14
#> 7  1.3     multi_stems       <NA>       001_1 2018.05.14
#> 8  1.3     multi_stems       <NA>       002_1 2018.05.13
#> 9  1.3 secondary_stems       <NA>       007_2 2018.05.14
#> 10 1.3 secondary_stems          1       007_3 2018.05.14
#> 11 1.3 secondary_stems          2       007_4 2018.05.14
#> 12 1.3 secondary_stems          3       007_5 2018.05.14
#> 13 1.3 secondary_stems       <NA>       004_2 2018.05.14
#> 14 1.3 secondary_stems          1       004_3 2018.05.14
#> 15 1.3 secondary_stems          2       004_4 2018.05.14
#> 16 1.3 secondary_stems          3       004_5 2018.05.14
#> 17 1.3 secondary_stems          4       004_6 2018.05.14
#> 18 1.3 secondary_stems          5       004_7 2018.05.14
#> 19 1.3 secondary_stems          6       004_8 2018.05.14
#> 20 1.3 secondary_stems       <NA>       007_2 2018.05.14
#> 21 1.3 secondary_stems          1       007_3 2018.05.14
#> 22 1.3 secondary_stems          2       007_4 2018.05.14
#> 23 1.3 secondary_stems          3       007_5 2018.05.14
#> 24 1.3 secondary_stems       <NA>       008_2 2018.05.14
#> 25 1.3 secondary_stems       <NA>       010_2 2018.05.14
#> 26 1.3 secondary_stems       <NA>       012_2 2018.05.14
#> 27 1.3 secondary_stems       <NA>       001_2 2018.05.14
#> 28 1.3 secondary_stems          1       001_3 2018.05.14
#> 29 1.3 secondary_stems       <NA>       002_2 2018.05.13
#> 30 1.3    single_stems       <NA>       006_1 2018.05.14
#> 31 1.3    single_stems       <NA>       008_1 2018.05.14
#> 32 1.3    single_stems       <NA>       009_1 2018.05.14
#> 33 1.3    single_stems       <NA>       010_1 2018.05.14
#> 34 1.3    single_stems       <NA>       001_1 2018.05.14
#> 35 1.3    single_stems       <NA>       002_1 2018.05.14
#> 36 1.3    single_stems       <NA>       003_1 2018.05.14
#> 37 1.3    single_stems       <NA>       005_1 2018.05.14
#> 38 1.3    single_stems       <NA>       006_1 2018.05.14
#> 39 1.3    single_stems       <NA>       009_1 2018.05.14
#> 40 1.3    single_stems       <NA>       011_1 2018.05.14
#> 41 1.3    single_stems       <NA>       002_1 2018.05.14
#> 42 1.3    single_stems       <NA>       003_1 2018.05.14
#> 43 1.3    single_stems       <NA>       004_1 2018.05.14
#> 44 1.3    single_stems       <NA>       001_1       <NA>
#> 45 1.3    single_stems       <NA>       002_1       <NA>
#> 46 1.3    single_stems       <NA>       001_1 2018.05.13
#> 47 1.3    single_stems       <NA>       003_1 2018.05.13
#> 48 1.3    single_stems       <NA>       004_1 2018.05.13
#> 49 1.3    single_stems       <NA>       005_1 2018.05.13
#> 

# Output excel
xl_sheets_to_xl(input_dir, output_dir, first_census = TRUE)

#> Warning: `.x` has no dataframe new_secondary_stems

#> Warning: `.x` has no dataframe recruits

# Read back
filename <- path(output_dir, "census.xlsx")
out <- read_excel(filename)
str(out, give.attr = FALSE)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	49 obs. of  15 variables:
#>  $ submission_id: chr  "c01b37d5-2c30-4086-b4a1-78327dfa0a9a" "efd8e991-2e34-40c8-b503-44def6e542d9" "efd8e991-2e34-40c8-b503-44def6e542d9" "efd8e991-2e34-40c8-b503-44def6e542d9" ...
#>  $ quadrat      : chr  "0816" "0916" "0916" "0916" ...
#>  $ tag          : chr  "007" "004" "007" "008" ...
#>  $ stem_tag     : chr  "1" "1" "1" "1" ...
#>  $ species      : chr  "American Basswood/Linden" "American Basswood/Linden" "Ironwood" "Ironwood" ...
#>  $ species_code : chr  "TILIAM" "TILIAM" "OSTRVI" "OSTRVI" ...
#>  $ dbh          : chr  "51.2" "31.6" "10.2" "2.7" ...
#>  $ status       : chr  "LI" "LI" "LI" "LI" ...
#>  $ codes        : chr  "M" "M" "M" "M;L" ...
#>  $ notes        : chr  NA NA NA "In brush pile" ...
#>  $ pom          : chr  "1.3" "1.3" "1.3" "1.3" ...
#>  $ sheet        : chr  "multi_stems" "multi_stems" "multi_stems" "multi_stems" ...
#>  $ section_id   : chr  NA NA NA NA ...
#>  $ unique_stem  : chr  "007_1" "004_1" "007_1" "008_1" ...
#>  $ date         : chr  "2018.05.14" "2018.05.14" "2018.05.14" "2018.05.14" ...