Combine spreadsheets from excel workbooks and output common data structures.

These functions combine spreadsheets from excel workbooks into common data structures. xl_sheets_to_csv() and xl_sheets_to_xl() write a .csv or excel (.xlsx) file per workbook -- combining all spreadsheets. xl_sheets_to_df outputs a list where each dataframes combines all spreadsheeets of a workbook.

xl_sheets_to_csv(input_dir, output_dir = "./")

xl_sheets_to_xl(input_dir, output_dir = "./")

xl_sheets_to_df(input_dir)

Arguments

input_dir
String giving the directory containing the excel workbooks to read from.

output_dir
String giving the directory where to write .csv files to.

Value

Writes one .csv file for each workbook.

Details

This is a rigid function with a very specific goal: To process data from a specific sampling software -- FastField. Specifically, this is what this function does:

Reads each spreadsheet from each workbook and map it to a dataframe.

Lowercases and links the names of each dataframe.

Keeps only these dataframes: (1) original_stems; (2) new_secondary_stems; and (3).

Dates the data by submission_id (date comes from the spreadsheet root).

Lowercases and links the names of each dataframe-variable.

Drops fake stems.

Output a common data structure of your choice.

Examples

library(fs)

# Path to the folder I want to read excel files from
input_dir <- dirname(example_path("two_files/new_stem_1.xlsx"))
input_dir

#> [1] "C:/Users/LeporeM/Dropbox/git_repos/fgeo.tool/inst/extdata/two_files"

# Files I want to read
dir(input_dir, pattern = "xlsx")

#> [1] "new_stem_1.xlsx" "new_stem_2.xlsx"

# Path to the folder I want to write .csv files to
output_dir <- tempdir()

# Do the work
xl_sheets_to_csv(input_dir, output_dir)

# Confirm
path_file(dir_ls(output_dir, regexp = "new_stem.*csv$"))

#> new_stem_1.csv new_stem_2.csv